==> 01_read_and_merge.Rmd <==
---
title: "MB1 CDI Follow-up Data Reading and Merge"
author: "The ManyBabies Analysis Team"
date: '`r format(Sys.time(), "%a %b %d %X %Y")`'
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: yes
---

# Intro

This script is the first preprocessing file for the primary ManyBabies 1 CDI Follow-up dataset. The goal is to get everything into a single datafile that can be used for subsequent analyses.

Data import functions are factored into a helper functions file.

```{r setup, message=FALSE, warning=FALSE, error=FALSE}
source("helper/preprocessing.R")
knitr::opts_chunk$set(cache = TRUE)

library(sessioninfo)
session_info(pkgs = "!attached", to_file = "01_session_log.txt")
```

# Data Import

## CDI Data Import

We first import the newly collected CDI data and useful lab information. Original data file names did not always convey encoding type (e.g. `.txt` for tab separated data), hence those were first manually renamed to facilitate automated data reading.

Other notable issues at data import were:

* a misnamed column in `infantlanglab-utk.tsv`, `CDI_agerange` instead of `CDI-agerange`, corrected by hand;
* a misnamed column in `chosunbaby-cdi.csv`, `vocab_nwords(=raw score)` instead of `vocab_nwords`, corrected by hand;
* a misnamed column in `infantstudies-ubc.tsv`, `subject_id` instead of `subid`, corrected by hand;
* extra non-needed columns in a few files, ignored by specifying by hand the columns and column types for import (see helper function);
* standardized scores from UK reported as numerical ranges instead of numbers. Not used in the current analyses but kept in a separate column as characters and in an unprocessed data file
(`data/cdi/unprocessed_merge.csv`); further, standardized score reported as `<1` were treated as zeros rather than NA;
* inconsistencies in the coding of `CDI.agerange`, sometimes coded as a ages in months rather than age groups (18, 24), recoded to 18 if younger than 21 months or 24 if older than 21 months;
* inconsistencies in the use of upper/lower case for the `webCDI` level of `CDI.form`.
* during data analysis, an error in age group assignment in `bllmanitoba` was discovered and corrected by hand
* Note: Some issues in updating 00_unprocessed_cdi_merged.csv. If you run into this problem, delete the file and have the code recreate it from scratch rather than updating. (There may be a more elegant solution to this problem.)

```{r cdi_import, message=FALSE, warning=FALSE}
# Lab information
data.labs <- read_labs("data/labs.csv")
# Participants data
data.cdi <- list.files("data/cdi_raw/") %>%
  map_dfr(read_cdi, save_unprocessed = !file.exists("data/00_unprocessed_cdi_merged.csv")) %>%
  drop_na(labid) %>%
  mutate(
    subid_unique = paste(labid, subid, sep = ":"),
    standardized.score.CDI.num = str_replace(standardized.score.CDI, "<1", "0"),
    CDI.agerange = case_when(
      !(CDI.agerange %in% c("18", "24")) ~
        if_else(CDI.agedays < 21 * 365.25 / 12, "18", "24"),
      T ~ as.character(CDI.agerange)
    ),
    CDI.form = str_replace(CDI.form, "webcdi", "webCDI"),
    CDI.error = (CDI.error == "error")
  ) %>%
  mutate_at(
    vars(subid, CDI.form, CDI.agerange),
    as_factor
  ) %>%
  left_join(data.labs) %>%
  mutate_at(vars(labid), as_factor) %>%
  mutate_at(vars(standardized.score.CDI.num), as.numeric)

# Check for labs that signed up but didn't send data:
missing_labs <- setdiff(
  levels(data.labs$labid),
  levels(data.cdi$labid)
) %>%
  str_sort()
```

At the moment, the following labs signed up but didn't send data (or possibly sent it since last data download): `r missing_labs`.

We expect to have no more than one data point per unique participant per age range (18 or 24 months):

```{r cdi_check, message=FALSE, warning=FALSE}
# Count data points, print cases were we have more than expected
data.cdi %>%
  count(subid_unique, CDI.agerange) %>%
  subset(n != 1)

# Only a few issues, let's check if they are subid typos or duplicate entries
data.cdi %>%
  group_by(subid_unique, CDI.agerange) %>%
  summarise(n = n(), n_unique = n_distinct(vocab_nwords)) %>%
  subset(n != 1)
```

Cases in which we have two data points but a single unique value for `vocab_nwords` are pure duplicate entries.

Checking by hand, we realise that the two duplicate data points for `babylablmu` result from a first entry with an error, and a second valid entry. The duplicate point for `unlv-musiclab` results from a similar error.

However, the duplicate for `infantcogubc` seems to be constituted of two distinct measures with no indication of an error that would allow us to discard one of them. This was checked with the lab, this was one participants that filled in the web questionnaire twice; the correct entry was the one where `vocab_nwords == 470` (see [issue #7](https://github.com/manybabies/mb1-cdi-followup/issues/7)).

Although the scope of this script is not to exclude data points, we remove the pure duplicates and `babylablmu`, `unlv-musiclab`, and `infantcogubc` "error duplicates" here while we know about them:

```{r cdi_fix_babylablmu, message=FALSE, warning=FALSE}
data.cdi <- data.cdi %>%
  subset(!(subid_unique %in% c("babylab-lmu:35", "babylab-lmu:76", "unlv-musiclab:MB021") &
    CDI.error %in% c("error", TRUE))) %>%
  subset(!(subid_unique == "InfantCog-UBC:tp13" & vocab_nwords != 470)) %>%
  unique()
```

## IDS Preference Data Import

We then import the (pre-processed) data from the original Many Babies study on IDS preference.

```{r ids_import, message=FALSE, warning=FALSE, error=FALSE}
# Import IDS data, rename variables, change column types
data.ids <- read_csv("data/ids_preference_trial.csv") %>%
  rename(labid = lab) %>%
  mutate_at(vars(labid, subid), as_factor) %>%
  mutate(gender = str_sub(gender, 1, 1)) %>%
  mutate(gender = case_when(
    gender == "0" ~ NA_character_,
    TRUE ~ gender
  )) # Some data coded as "(FE)MALE" instead of F/M and zero
```

## Compatibility Check

We can now start checking that the merge is possible.

### Coding of `labid`

We first check that the coding of `labid` is consistent between the original IDS data and the new CDI data.

```{r labid_check}
# Check for labid present in CDI but not IDS (i.e. differently named), and vice versa
changed_labid <- setdiff(
  data.cdi$labid,
  data.ids$labid
)
unused_labid <- setdiff(
  data.ids$labid,
  data.cdi$labid
)
# By hand, we see that the names seem to be the same but with "-" or "_" added in data.cdi
# Only exception: leeds-lcdu in data.cdi should be lcduleeds
data.cdi <- data.cdi %>%
  mutate(labid = case_when(
    labid == "leeds-lcdu" ~ "lcduleeds",
    labid == "InfantCog-UBC" ~ "infantcogubc",
    labid == "princeton-babylab" ~ "babylabprinceton",
    labid == "um_babylanguagelab" ~ "bllumanitoba",
    T ~ str_remove_all(labid, "-|_")
  )) %>%
  mutate_at(vars(labid), as_factor)
# We check again for safety
changed_labid <- setdiff(
  data.cdi$labid,
  data.ids$labid
)
```

### Coding of `subid`

Now that we have consistent coding for `labid`, we can update the new `data.cdi$subid_unique` and check for consistency at that level (if there is a new `subid_unique` that does not correspond to an old one, then the corresponding `subid` is faulty or the participant is missing from `data.ids`).

```{r subid_check}
# Update subid_unique in CDI
data.cdi <- data.cdi %>%
  mutate(subid_unique = paste(labid, subid, sep = ":"))

# Check for labid present in CDI but not IDS (i.e. differently named)
changed_subid <- setdiff(
  data.cdi$subid_unique,
  data.ids$subid_unique
)

# Many subids to check, let's extract the labs with issues
changed_subid_labs <- changed_subid %>%
  str_match("[a-z]+") %>%
  unique()

# Too many lab, let's print for each lab the proposed subids (CDI) and the possible subids (IDS)
unused_subid <- setdiff(
  data.ids$subid_unique,
  data.cdi$subid_unique
)
check_subid <- changed_subid_labs %>%
  map(~ list(
    CDI = str_subset(changed_subid, .x) %>% str_sort(),
    IDS = str_subset(unused_subid, .x) %>% str_sort()
  )) %>%
  setNames(changed_subid_labs)

# Fix formatting issues, update subid_unique
data.cdi <- data.cdi %>%
  mutate_at(vars(subid), as.character) %>%
  mutate(
    subid = case_when(
      labid == "babylabbrookes" ~ str_replace(subid, "brookes-", "mb"),
      labid == "bcrlunlv" ~ paste0("mb", str_pad(subid, 4, pad = 0)),
      labid == "lscppsl" ~ paste0(subid, "lscppsl"),
      labid == "babylabprinceton" ~ str_replace(
        subid, "MB",
        "manybabies"
      ),
      labid == "unlvmusiclab" ~ str_replace(subid, "MB", "run"),
      labid %in% changed_subid_labs ~
        str_remove_all(subid, "-|_") %>% str_to_lower(),
      T ~ subid
    ),
    subid_unique = paste(labid, subid, sep = ":")
  ) %>%
  mutate_at(vars(subid, subid_unique), as_factor)
# Check again for safety
changed_subid <- setdiff(
  data.cdi$subid_unique,
  data.ids$subid_unique
)
unused_subid <- setdiff(
  data.ids$subid_unique,
  data.cdi$subid_unique
)
changed_subid_labs <- changed_subid %>%
  str_match("[a-z]+") %>%
  unique()
check_subid <- changed_subid_labs %>%
  map(~ list(
    CDI = str_subset(changed_subid, .x) %>% str_sort(),
    IDS = str_subset(unused_subid, .x) %>% str_sort()
  )) %>%
  setNames(changed_subid_labs)
check_subid.tibble <- check_subid %>%
  map_dfr(~ tibble(CDI = list(.x[["CDI"]]), IDS = list(.x[["IDS"]])), .id = "labid") %>%
  gather(key = "dataset", value = "subids", -labid) %>%
  unnest(subids) %>%
  mutate(subids = str_remove(subids, "[a-z]+:")) %>%
  mutate_all(as_factor)
```

Even after correcting formatting issues in subids, there are `r sum(check_subid.tibble$dataset == "CDI")` subids in the CDI data that do not have a match in the preprocessed IDS data. This is most likely due to the fact that some infants were excluded from the IDS analysis but still had their CDI scores recorded. We check that this is the case below.

```{r subid_exclusion_check, message=FALSE, warning=FALSE}
# Import subids from all participants in IDS (before exculsion) for labs that provided CDI data
data.ids.full.subids <- read_csv("data/ids_full.csv",
  col_types = cols_only(
    lab = col_character(),
    subid_unique = col_character()
  )
) %>%
  rename(labid = lab) %>%
  subset(labid %in% levels(data.cdi$labid))

# Compare to subids in CDI
changed_subid.full <- setdiff(
  data.cdi$subid_unique,
  data.ids.full.subids$subid_unique
)
unused_subid.full <- setdiff(
  data.ids.full.subids$subid_unique,
  data.cdi$subid_unique
)
changed_subid_labs.full <- changed_subid.full %>%
  str_match("[a-z]+") %>%
  unique()
check_subid.full <- changed_subid_labs.full %>%
  map(~ list(
    CDI = str_subset(changed_subid.full, .x) %>% str_sort(),
    IDS = str_subset(unused_subid.full, .x) %>% str_sort()
  )) %>%
  setNames(changed_subid_labs.full)
check_subid.tibble.full <- check_subid.full %>%
  map_dfr(~ tibble(CDI = list(.x[["CDI"]]), IDS = list(.x[["IDS"]])), .id = "labid") %>%
  gather(key = "dataset", value = "subids", -labid) %>%
  unnest(subids) %>%
  mutate(subids = str_remove(subids, "[a-z]+:")) %>%
  mutate_all(as_factor)

# Save problematic labs' unmatched subids
check_subid.tibble.full %>%
  nest(data = c(dataset, subids)) %>%
  pwalk(~ write_csv(.y, paste0("data/subids/", .x, ".csv")))
```

Here we see that `r sum(check_subid.tibble.full$dataset == "CDI")` subids from the CDI data still have no match even in the unprocessed IDS data. Those might have resulted from typos when entering data, and thus need to be handchecked with the corresponding labs.

* For `weltentdeckerzurich`, the `r check_subid.tibble.full %>% filter(dataset=="CDI"&labid=="weltentdeckerzurich") %>%count()` unmatched CDI subids were confirmed to be infants that were not included in the IDS analysis (see [issue #2](https://github.com/manybabies/mb1-cdi-followup/issues/2)).
* Similarly for `babylabbrookes`, the `r check_subid.tibble.full %>% filter(dataset=="CDI"&labid=="babylabbrookes") %>%count()` unmatched CDI subids were infants that were not included in the IDS analysis (see [issue #3](https://github.com/manybabies/mb1-cdi-followup/issues/3)). Notably participant `mb43` in our dataset should have been `mb143`, but this does not matter given that we do not keep their data.
* For `chosunbaby`, the only remaining issue was a typo, corrected below, where participant `h71y23` should be `h71o23`, (see [issue #4](https://github.com/manybabies/mb1-cdi-followup/issues/4)).
* For `infantlanglabutk`, the `r check_subid.tibble.full %>% filter(dataset=="CDI"&labid=="infantlanglabutk") %>%count()` unmatched CDI subids were infants for whom there was no IDS data available (see [issue #5](https://github.com/manybabies/mb1-cdi-followup/issues/5)).
* For `unlvmusiclab` the only mismatch stems from the participant having been a pilot infant for the IDS analysis. We correct its subid here for consistency, but this participant will be exclude in script 02 (see [issue #6](https://github.com/manybabies/mb1-cdi-followup/issues/6)). 

```{r subid_handcheck, message=FALSE, warning=FALSE}
data.cdi <- data.cdi %>%
  mutate(
    subid = case_when(
      labid == "chosunbaby" ~ str_replace(subid, "h71y23", "h71o23"),
      labid == "unlvmusiclab" ~ str_replace(subid, "run001", "run001pilot"),
      T ~ as.character(subid)
    ),
    subid_unique = paste(labid, subid, sep = ":")
  ) %>%
  mutate_at(vars(subid, subid_unique), as_factor)
```

# Merge

Finally we merge the IDS preference data into the CDI data, so that only IDS preference data from infants whose CDI score was recorded is kept in the final dataset.

```{r data_merge, message=FALSE, warning=FALSE, error=FALSE}
data.merged <- left_join(data.cdi, data.ids)
write_csv(data.merged, "data/01_processed_merged.csv")
```
