==> 02_exclusion_and_new_variables.Rmd <==
---
title: "MB1 CDI Follow-up Exclusions and Data Manipulations"
author: "The ManyBabies Analysis Team"
date: '`r format(Sys.time(), "%a %b %d %X %Y")`'
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: yes
---

# Intro

This script implements and documents exclusion criteria. We first read the full merged dataset produced by the first script `01_read_and_merge.Rmd`.

```{r setup, message=FALSE, warning=FALSE}
library(tidyverse)
knitr::opts_chunk$set(cache = TRUE)
```

```{r session_infos}
# Provides a session log for the packages used in this .rmd
library(sessioninfo)
session_info(pkgs = "!attached", to_file = "02_session_log.txt")
```

```{r col_types, message=FALSE, warning=FALSE}
# Define column types
col_types <- cols(
  labid = col_factor(),
  subid = col_factor(),
  CDI.form = col_factor(),
  CDI.agerange = col_factor(),
  CDI.agedays = col_integer(),
  vocab_nwords = col_integer(),
  standardized.score.CDI = col_character(),
  standardized.score.CDI.num = col_number(),
  CDI.error = col_logical(),
  Notes = col_character(),
  language = col_factor(),
  language_zone = col_factor(),
  subid_unique = col_factor(),
  trial_order = col_factor(),
  trial_num = col_factor(),
  trial_type = col_factor(),
  stimulus_num = col_factor(),
  method = col_factor(),
  age_days = col_integer(),
  age_mo = col_number(),
  age_group = col_factor(),
  nae = col_logical(),
  gender = col_factor(),
  second_session = col_logical(),
  looking_time = col_number(),
  missing = col_logical()
)
```

# Data Import
```{r import_data}
# Read data
df <- read_csv("data/01_processed_merged.csv", col_types = col_types)
```

The total number of rows of data in our merged dataset is `r count(df)`.

# Exclusions

Before excluding participants based on the exclusion criteria reported in our manuscript, we exclude rows for which we don't have a raw score.

```{r remove_na}
exclusions.na <- df %>%
  drop_na(vocab_nwords) %>%
  droplevels()
```

The total number of rows of data in our merged dataset is reduced to `r count(exclusions.na)`, meaning that `r count(df) - count(exclusions.na)` rows of data were excluded. We are left with `r length(levels(exclusions.na$subid_unique))` participants out of `r length(levels(df$subid_unique))`, meaning that `r length(levels(df$subid_unique)) - length(levels(exclusions.na$subid_unique))` participants were excluded due to missing data.

## Infants Outside Age Range

We first exclude all infants who are outside the age range described in the data collection protocol. These are:

* for the North American sample, 16 to 20 months and 22 to 26 months for the 18mo and 24mo age groups respectively;
* for all other participants sample, 17.5 to 18.5 months and 23.5 to 24.5 months.

```{r age_range}
# Age ranges
exclusions.age <- exclusions.na %>%
  mutate(
    CDI.agemin = map2_dbl(
      language_zone, as.integer(as.character(CDI.agerange)),
      ~ round((.y - if_else(.x == "NAE", 2, 0.5)) * 365.25 / 12)
    ),
    CDI.agemax = map2_dbl(
      language_zone, as.integer(as.character(CDI.agerange)),
      ~ round((.y + if_else(.x == "NAE", 2, 0.5)) * 365.25 / 12)
    )
  ) %>%
  subset(!(CDI.agedays < CDI.agemin | CDI.agedays > CDI.agemax)) %>%
  droplevels()
```

After the age exclusion, we are left with `r length(levels(exclusions.age$subid_unique))` participants out of `r length(levels(exclusions.na$subid_unique))`, meaning that `r length(levels(exclusions.na$subid_unique)) - length(levels(exclusions.age$subid_unique))` were excluded due to falling outside of the planned age range.

## Incomplete IDS/ADS pairs and Infants Excluded from MB1 IDS

For analysis purposes, we consider pairs of ADS/IDS trials, and as such exclude any pairs in which one of the trials is missing. Participants without any useable trials after these exclusions are excluded from further analysis.

Labs sometimes recorded CDI scores regardless of the inclusion of the IDS data in the original study. Those infants and only those will have `NA` values for columns such as `trial_order`, `trial_num`, `trial_type`. This also excludes these infants.


```{r trial_pairs}
exclusions.pairs <- exclusions.age %>%
  subset(!missing, select = -missing) %>%
  # We have to add vocab_nwords as some infants have multiple data points for filling the CDI twice (processed further down)
  group_by(subid_unique, CDI.agerange, vocab_nwords, stimulus_num) %>%
  mutate(pair_complete = (n() == 2)) %>%
  ungroup() %>%
  subset(pair_complete,
    select = -pair_complete
  ) %>%
  droplevels()
```

Our dataset is now reduced to `r length(levels(exclusions.pairs$subid_unique))` from  `r length(levels(exclusions.age$subid_unique))` participants, meaning that `r length(levels(exclusions.age$subid_unique)) - length(levels(exclusions.pairs$subid_unique))` participants were excluded due to insufficient IDS preference data.

## Reported errors

Other than the exclusion criteria reported above, labs reported errors relating to the CDI data gathering. We treat those errors here by hand.

```{r cdi_errors_check}
# Check the different kinds of errors with notes added
errors_check <- exclusions.pairs %>%
  subset(CDI.error) %>%
  select(labid:language_zone) %>%
  mutate_at(vars(Notes), as_factor)
# Are there any errors for which we don't have notes?
(anyNA(errors_check$Notes))
# What kind of errors do we have?
(levels(errors_check$Notes))
```

We have a limited number of error types to process, and no participants that had an error with no attached note. The "aged out" error was due to differences in how indivual labs converted age in days to age in months and how we did it in this script, and can be ignored. For the last type of error, "`r last(levels(errors_check$Notes))`, there does not seem to be an issue with the data itself, and this error can therefore be ignored. Only the first error "`r first(levels(errors_check$notes))` appears to be a true error and is excluded.

```{r cdi_errors_exclusions}
exclusions.errors <- exclusions.pairs %>%
  filter(is.na(Notes) | Notes != "Opened and did not complete the questionnaire") %>%
  droplevels()
```


Our dataset is now reduced from `r length(levels(exclusions.pairs$subid_unique))` participants to `r length(levels(exclusions.errors$subid_unique))`, meaning that `r  length(levels(exclusions.pairs$subid_unique))-length(levels(exclusions.errors$subid_unique))` participants were excluded due to a reported error, namely that the data were incomplete. It is worth noting that this is likely a significant underestimate of the number of participants who started but did not complete the questionnaire as most labs did not specifically report these data, or they were excluded due to missing data at a prior exclusion stage.

## Final dataset

We now have a full dataset with only participants that matched the inclusion criteria defined in our manuscript. We save this dataset, and move on to computing new variables that are needed for our analysis plan.

```{r save_exclusions}
write_csv(exclusions.errors, "data/02a_exclusions.csv")
```

# New Variables

## `IDS_pref`

We want to create a new variable for the standardized mean preference for IDS `z_IDS_pref`. To do so, we first need to compute the IDS preference for each pair of trials for each infants. We also need to compute the average looking time across both ADS and IDS trials.

```{r ids_preference}
df.final <- exclusions.errors %>%
  select(-trial_num) %>% # Not relevant, we only want the pair number in stimulus_num
  pivot_wider(names_from = trial_type, values_from = looking_time) %>%
  mutate(
    IDS_pref_pair = IDS - ADS,
    mean_LT_pair = (IDS + ADS) / 2
  )
```

We then need to compute the mean preference for IDS for each infant, and the mean looking time across trials for each infant, to finally get our standardized preference score for IDS.

```{r z_ids_preference}
df.final <- df.final %>%
  group_by(subid_unique) %>%
  mutate(
    mean_IDS_pref = mean(IDS_pref_pair),
    mean_LT = mean(mean_LT_pair),
    IDS_pref = mean_IDS_pref / mean_LT
  ) %>%
  ungroup() %>%
  select(-c(stimulus_num, IDS, ADS, IDS_pref_pair, mean_LT_pair, mean_IDS_pref, mean_LT)) %>%
  unique()
```


```{r}
df.final <- df.final %>%
  mutate(z.IDS_pref = as.numeric(scale(IDS_pref, center = T, scale = F)))
```

## `CDI.nwords`, `CDI.prop`

For our 3rd research question, we need to compute the proportional CDI scores as the raw score (number of words produced) divided by the total number of words in the CDI used.


```{r CDI.prop}
CDI.nwords.list <- tribble(
  ~CDI.form, ~CDI.nwords,
  "O-CDI", 416,
  "webCDI", 680
)

df.final <- df.final %>%
  left_join(CDI.nwords.list) %>%
  mutate_at(vars(CDI.form), as_factor) %>%
  mutate(CDI.prop = vocab_nwords / CDI.nwords)
```

## `z_age_months`, `CDI.z_age_months`, `CDI.grouped.z_age_months`

A common way to get better computational models both in terms of estimate precision and convergence is to center and scale numerical variables, particularly ones that have relatively large values and/or large spreads. In our case, age in days are such variables. However, for better interpretation and generalisation of the results, we choose to center those values, but instead of scaling them we simply convert them to ages in months. Here, we further create variables to center `CDI.agedays` either overall or for each age group defined by `CDI.agerange`.

```{r z_age_days}
df.final <- df.final %>%
  mutate(
    z_age_months = as.numeric(scale(age_days, scale = F)) / 365.25 * 12,
    CDI.z_age_months = as.numeric(scale(CDI.agedays, scale = F)) / 365.25 * 12
  ) %>%
  group_by(CDI.agerange) %>%
  mutate(CDI.grouped.z_age_months = as.numeric(scale(CDI.agedays, scale = F)) / 365.25 * 12) %>%
  ungroup()
```

## `z_standardized_CDI`, `z_vocab_nwords`

We discussed above the interest of scaling numerical variables. In our case, our dependant variables are also numerical with high values and wide spreads, so scaling and centring them can sensibly improve model convergence and model fit.

```{r z_CDI_scores}
df.final <- df.final %>%
  mutate(
    z_standardized_CDI = as.numeric(scale(standardized.score.CDI.num)),
    z_vocab_nwords = as.numeric(scale(vocab_nwords))
  )

# remove extra dimension from z score variables (value of 1 in second column prevented write_csv in next chunk)
df.final$z_age_months <- drop(df.final$z_age_months)
df.final$CDI.z_age_months <- drop(df.final$CDI.z_age_months)
df.final$CDI.grouped.z_age_months <- drop(df.final$CDI.grouped.z_age_months)
df.final$z_standardized_CDI <- drop(df.final$z_standardized_CDI)
df.final$z_vocab_nwords <- drop(df.final$z_vocab_nwords)
df.final$z.IDS_pref <- drop(df.final$z.IDS_pref)
```

# Conclusion

In the end, our dataset consists of `r length(levels(df.final$subid_unique))` infants and `r nrow(df.final)` CDI scores at 18 or 24 months.


```{r data_summary}
write_csv(df.final, "data/02b_processed.csv")
df.final %>% count(CDI.agerange)
```

Some additional tables for sample reporting are below.

```{r participant reporting}
lab.table <- df.final %>%
  group_by(labid) %>%
  summarise(
    min.mb1.age = min(age_days), max.mb1.age = max(age_days),
    n.18months = sum(CDI.agerange == "18"),
    n.24months = sum(CDI.agerange == "24"),
    n.m.18months = sum(gender == "M" & CDI.agerange == "18"),
    n.m.24months = sum(gender == "M" & CDI.agerange == "24")
  ) %>%
  arrange(labid)
knitr::kable(lab.table, caption = "Data for Table 1 in manuscript")
```
