==> 03_confirmatory_analysis.Rmd <==
---
title: "MB1 CDI Follow-up Confirmatory Analysis"
author: "The ManyBabies Analysis Team"
date: '`r format(Sys.time(), "%a %b %d %X %Y")`'
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: yes
---

# Introduction

In this script, we go through all the pre-registered proposed analyses. As a reminder, the main research questions where as follows:

1. To what extent does infants' preference for IDS as measured in a laboratory setting predict their vocabulary at 18 and 24 months?
2. Does the relation between IDS preference and vocabulary size change over development?
3. Are there systematic differences in the strength of this relationship across the language communities in our sample?

We first present the main "sample theory based" analyses (also known as frequentist), separately on the North American and UK samples in parallel to answer our first two research questions, then on the total dataset to answer our third research question. We then provide additional Bayesian statistics where a null effect was found, as specified in the pre-registration.

```{r setup, message=FALSE, warning=FALSE}
# Library imports, general settings ==============
library(tidyverse); library(egg)
library(knitr)
library(lme4); library(lmerTest); library(simr)
# As in our discussion with Mike, we will use lmerTest for calculating p values
# in the mixed-effects models (noted as a deviation) 
library(brms); library(rstan)
library(future)
plan(multicore, workers = 8) # Swap to multiprocess if running from Rstudio
library(lattice)
library(effects)
library(sjPlot)
library(robustlmm)
library(car)

# Load model comparison functions
source("helper/lrtests.R")

# Deal with package priority issues
select <- dplyr::select

theme_set(theme_bw(base_size = 10))
options("future" = T)
#knitr::opts_chunk$set(cache = TRUE)

print(sessionInfo()) #listing all info about R and packages info

# Read data ======================================
col_types <- cols(
  labid = col_factor(),
  subid = col_factor(),
  subid_unique = col_factor(),
  CDI.form = col_factor(),
  CDI.nwords = col_integer(),
  CDI.prop = col_number(),
  CDI.agerange = col_factor(),
  CDI.agedays = col_integer(),
  CDI.agemin = col_integer(),
  CDI.agemax = col_integer(),
  vocab_nwords = col_integer(),
  standardized.score.CDI = col_character(),
  standardized.score.CDI.num = col_number(),
  IDS_pref = col_number(),
  language = col_factor(),
  language_zone = col_factor(),
  CDI.error = col_logical(),
  Notes = col_character(),
  trial_order = col_factor(),
  method = col_factor(),
  age_days = col_integer(),
  age_mo = col_number(),
  age_group = col_factor(),
  nae = col_logical(),
  gender = col_factor(),
  second_session = col_logical()
)
data.total <- read_csv("data/02b_processed.csv", col_types = col_types)
# TODO: add saved results
```

Before moving on with the analysis, we have to ready the data by (a) checking for colinearity between `z_age_months` and `CDI.z_age_months` and correcting this if necessary, and (b) setting up the contrasts described in our data analysis.

## Colinearity check

First, we run a Kappa test on the possibility of colinearity between `z_age_months` and `CDI.z_age_months`.

```{r colinearity_check}
# Run kappa test
k.age_months <- model.matrix(~ z_age_months + CDI.z_age_months, data = data.total) %>%
  kappa(exact = T)
```

With a value of `r k.age_months`, we do not have a colinearity issue and can proceed with the data analysis as planned (The criteria of indicating colinearity is that kappa > 10). 

## Contrast Setups

We need `gender` as an effect-coded factor, and `method` as a deviation-coded factor. This is achieved in R by using the `contr.sum()` function with the number of levels for each factor. Notably, when subsetting the UK sample, only two levels of `method` out of the three in total were left. 

```{r contrasts}
# Set contrasts on the total dataset =============
contrasts(data.total$gender) <- contr.sum(2)
contrasts(data.total$method) <- contr.sum(3)

# Create sub-datasets, with contrasts ============
## NAE
data.nae <- data.total %>% subset(language_zone == "NAE") %>% droplevels()
contrasts(data.nae$gender) <- contr.sum(2)
contrasts(data.nae$method) <- contr.sum(3)

## UK (separate 18/24 months)
data.uk.18 <- data.total %>% subset(language_zone == "British" & CDI.agerange == 
                                "18") %>% droplevels()
contrasts(data.uk.18$gender) <- contr.sum(2)
contrasts(data.uk.18$method) <- contr.sum(2) #note that UK sample has only 2 levels, so sum of zero contrasts set to 2 levels

data.uk.24 <- data.total %>% subset(language_zone == "British" & CDI.agerange == 
                                "24") %>% droplevels()
contrasts(data.uk.24$gender) <- contr.sum(2)
contrasts(data.uk.24$method) <- contr.sum(2) #note that UK sample has only 2 levels, so sum of zero contrasts set to 2 levels



## Other
data.other <- data.total %>% subset(language_zone == "Other") %>% droplevels()
contrasts(data.other$gender) <- contr.sum(2)
contrasts(data.other$method) <- contr.sum(3)
```

# Descriptive Statistics

We first assess the amount of data we have overall per condition and their shape overall.

```{r desc_total}
data.total %>%
  group_by(language_zone, CDI.agerange, method, gender) %>%
  summarise(N = n(), age = mean(CDI.agedays), sd = sd(CDI.agedays)) %>%
  kable()
```
More detailed information about Descriptive Statistics
```{r}
#number of lab
data.total %>% 
  select(labid, language_zone) %>% 
  unique() %>% 
  group_by(language_zone) %>% 
  count()

data.total %>% 
  group_by(language_zone, CDI.agerange) %>% 
  summarize(N = n())

# age range in each age group and language zone
data.total %>% 
  select(subid, language_zone, CDI.agedays, CDI.agerange) %>% 
  unique() %>% 
  group_by(language_zone, CDI.agerange) %>% 
  summarize(age_min = (min(CDI.agedays)/365.25*12),
            age_max = (max(CDI.agedays)/365.25*12))

```

We then assess the data per lab in terms of sample size and CDI score (vocabulary size, for consistency between language zones).

```{r desc_by_lab}
by_lab <- data.total %>%
  group_by(labid, language_zone, CDI.agerange) %>%
  mutate(tested = n_distinct(subid_unique)) %>%
  select(labid, language_zone, CDI.agerange, tested, vocab_nwords) %>%
  nest(scores = vocab_nwords) %>%
  mutate(model = map(scores, ~ lm(vocab_nwords ~ 1, data = .x)),
         ci = map(model, confint)) %>%
  transmute(tested = tested,
            mean = map_dbl(model, ~ coefficients(.x)[[1]]),
            ci_lower = map_dbl(ci, 1),
            ci_upper = map_dbl(ci, 2)) %>%
  arrange(language_zone) %>%
  rownames_to_column()

# TODO: find a way to group by language zone?
ggplot(by_lab,
       aes(x = labid, colour = language_zone,
           y = mean, ymin = ci_lower, ymax = ci_upper)) + 
  geom_linerange() + 
  geom_point(aes(size = tested)) + 
  facet_grid(cols = vars(CDI.agerange), scales = "free") + coord_flip(ylim = c(0, 500)) +
  xlab("Lab") + ylab("Vocabulary size") +
  scale_colour_brewer(palette = "Dark2", name = "Language zone",
                      breaks = c("British", "NAE", "Other")) +
  scale_size_continuous(name = "N", breaks = function(x) c(min(x), mean(x), max(x))) +
  theme(legend.position = "bottom")
```

# Sample Theory Based Statistics

## Simple Correlation

First, we want to assess quickly if there is a direct correlation between IDS preference and CDI score, computing a Pearson#'s product-moment correlation. We use standardized CDI scores for the North American sample, and raw scores for the British sample. Since CDI grows with age, we run the British sample separately for 18 and 24 months.

```{r simple_correlation}
# Statistics =====================================
## North American Sample
test.pearson.nae <- cor.test(data.nae$IDS_pref,
                             data.nae$z_standardized_CDI,
                             alternative = "two.sided", method = "pearson")

test.pearson.nae

## UK Sample
test.pearson.uk.18 <- cor.test(data.uk.18$IDS_pref,
                            data.uk.18$z_vocab_nwords,
                            alternative = "two.sided", method = "pearson")

test.pearson.uk.18

test.pearson.uk.24 <- cor.test(data.uk.24$IDS_pref,
                            data.uk.24$z_vocab_nwords,
                            alternative = "two.sided", method = "pearson")

test.pearson.uk.24


```


Plots for correlation
```{r}
## North American Sample
### Get correlation value for annotation
cor_text <- "paste(italic(R)^2, \" =\")"
cor_value <- round(test.pearson.nae$estimate, 3)

### Build plot
plot.pearson.nae <- data.nae %>%
  ggplot(aes(x = IDS_pref,
             y = standardized.score.CDI.num)) +
  xlab("IDS preference") + ylab("Standardized CDI score") +
  geom_point() +
  geom_smooth(method = lm) +
  annotate("text", x = -.9, y = 51, parse = T, size = 4,
           label = paste(cor_text, cor_value, sep = "~"))

## UK Sample
cor_value <- round(test.pearson.uk$estimate, 3)
plot.pearson.uk <- data.uk %>%
  ggplot(aes(x = IDS_pref,
             y = vocab_nwords)) +
  xlab("IDS preference") + ylab("Vocabulary size (in words)") +
  geom_point() +
  geom_smooth(method = lm) +
  annotate("text", x = .8, y = 150, parse = T, size = 4,
           label = paste(cor_text, cor_value, sep = "~"))

# Global plot
plot.pearson <- ggarrange(plot.pearson.nae, plot.pearson.uk, ncol = 2)

plot.pearson

ggsave("plots/corr_plot.pdf", plot.pearson,
       units = "mm", width = 180, height = 100, dpi = 1000)

```

We see no obvious direct link between IDS prefernce and CDI score here. However, an effect might appear once we take into account various factors that might interact with IDS preference and/or CDI score. We can also first enhance these plots with information about the age group at which infants were tested (18- or 24-month-old), using vocabulary size to better compare the NAE and UK samples.

```{r plot_by_age}
plot.age_group <- data.total %>%
  subset(language_zone != "Other") %>%
  droplevels() %>%
  ggplot(aes(x = IDS_pref,
             y = vocab_nwords,
             colour = CDI.agerange)) +
  facet_wrap(vars(language_zone),
             labeller = as_labeller(c("British" = "UK samples",
                                      "NAE" = "North Amercian samples"))) +
  xlab("Standardized IDS prefence score") + ylab("Vocabulary size (in words)") + theme(legend.position = "top") +
  geom_point() +
  geom_smooth(method = lm) +
  scale_colour_brewer(palette = "Dark2", name = "Age group",
                      breaks = c("18", "24"),
                      labels = c("18mo", "24m"))
ggsave("plots/scatter_age.pdf", plot.age_group,
       units = "mm", width = 180, height = 100, dpi = 1000)
(plot.age_group)
```

## Mixed-Effects Model by Language Zone

Here, we run a mixed-effects model including only theoretically motivated effects, as described in the pre-registration. We start with the full model bellow, simplifying the random effects structure until it converges.

### NAE full model
```{r full_lmer}
# Run models =====================================
## NAE

data.nae$centered_IDS_pref <- scale(data.nae$IDS_pref, center = T, scale = F)

lmer.full.nae <- lmer(standardized.score.CDI.num ~ CDI.z_age_months + gender +
                        z_age_months + method + centered_IDS_pref +
                        centered_IDS_pref:method + centered_IDS_pref:CDI.z_age_months + centered_IDS_pref:z_age_months +
                        (1 | labid) + (1 | subid_unique),
                      data = data.nae)

summary(lmer.full.nae)

# robust_lmer.full.nae <- robustlmm::rlmer(standardized.score.CDI.num ~ CDI.z_age_months + gender +
#                         z_age_months + method + centered_IDS_pref +
#                         centered_IDS_pref:method + centered_IDS_pref:CDI.z_age_months + centered_IDS_pref:z_age_months +
#                         (1 | labid),
#                       data = data.nae)
# 
# 
# summary(robust_lmer.full.nae) #this model is used to see if we can meet some statistical assumption better but we decided to use the original model as the inferential statistical results are consistent

full.nae_pvalue <- anova(lmer.full.nae) %>% 
  as_tibble(rownames = "Parameter") #this gives us the Type III p values

# ==========
# Sequentially removed random effects:
# IDS_pref:z_age_months
# IDS_pref:CDI.z_age_months
# IDS_pref:method
# IDS_pref
# method
# z_age_months
# gender
# CDI.z_age_months
#==========

```

#### (Optional) Checking mixed-model assumptions. We will check the following:
1) Linearlity
2) Normality of the residuals
3) Homoscedasticity of residuals
4) No autocorrelation
5) No multicollinearity
```{r}
#First, check linearity
# data.nae$resid <- residuals(lmer.full.nae)
# 
# plot(data.nae$resid, data.nae$standardized.score.CDI)

#Second, check normality
plot_model(lmer.full.nae, type = 'diag') ## we do have right-skewed normality of residuals

#Third, check autocorrelation
re_run_lme.full.nae <- nlme::lme(standardized.score.CDI.num ~ CDI.z_age_months + gender +
                        z_age_months + method + centered_IDS_pref +
                        centered_IDS_pref:method + centered_IDS_pref:CDI.z_age_months + centered_IDS_pref:z_age_months, random = ~1 | labid,
                        method = "REML",
                      data = data.nae, na.action = na.exclude)

plot(nlme::ACF(re_run_lme.full.nae, resType = "normalized")) #there is no sign for autocorrelation
      
#Lastly, check multi-collinearity
car::vif(lmer.full.nae) #we do see a multicollineartiy for the IDS preference variable, even though we have centered the IDS preference score. It is probably related to the number the participating labs (as this is the group level that we are controlling) and how we entered interaction between IDS preference and other variables (that lack variability in the current sample). We need to keep IDS preference in the model as exploring the relationship between IDS preference and CDI score is the key research question in the paper.
```


### UK full model
```{r}
lmer.full.uk <- lmer(vocab_nwords ~ CDI.z_age_months + gender +
                       z_age_months + method + IDS_pref +
                       IDS_pref:method + IDS_pref:CDI.z_age_months + IDS_pref:z_age_months +
                       (1 | labid) + (1 | subid_unique),
                     data = data.uk) 

summary(lmer.full.uk)

full.uk_pvalue <- anova(lmer.full.uk) %>% 
  as_tibble(rownames = "Parameter") #this gives us the Type III p values

#==========
# Sequentially removed random effects:
# IDS_pref:z_age_months
# IDS_pref:CDI.z_age_months
# IDS_pref:method
# IDS_pref
# method
# z_age_months
# gender
# CDI.z_age_months

```
#### (Optional) Checking mixed-model assumptions. We will check the following:
1) Linearlity
2) Normality of the residuals
3) Homoscedasticity of residuals
4) No autocorrelation
5) No multicollinearity
```{r}
#First, check linearity. The plot looks linear
data.uk$resid <- residuals(lmer.full.uk)
 
plot(data.uk$resid, data.uk$vocab_nwords)

#Second, check normality
plot_model(lmer.full.uk, type = 'diag') ## we do have right-skewed normality of residuals

#Third, check autocorrelation
re_run_lme.full.uk <- nlme::lme(vocab_nwords ~ CDI.z_age_months + gender +
                       z_age_months + method + IDS_pref +
                       IDS_pref:method + IDS_pref:CDI.z_age_months + IDS_pref:z_age_months, random = ~1 | labid,
                        method = "REML",
                      data = data.nae, na.action = na.exclude)

plot(nlme::ACF(re_run_lme.full.uk, resType = "normalized")) #there is no sign for autocorrelation
      
#Lastly, check multi-collinearity
car::vif(lmer.full.uk) #no problem for multicollinearlity
```


We now want to check the statistical power of significant effects, and discard any models with significant effects that do not reach 80% power. This however leads to too many warnings of singularity issues on the model updates inherent to the `simr` power simulations, hence we cannot obtain satisfactory power estimates as pre-registered.

AST: Note that we don't have any IV(s) that turned out to be significant in the Full NAE model. So we won't run the power analysis check. For the UK full model, there are two statistically significant IV: CDI_age and gender. The post hoc power check suggested that we have high power in detecting the effect of CDI_age but not gender. Note that gender has a smaller effect size to begin with, so this may partially explain why we have less power in detecting it in the model. As there can be a number of different factors that determines the posthoc power, we decided not to remove gender in the model based on posthoc power analysis check. 

```{r pwr_lmer, eval = FALSE}

check_pwr_uk_cdi_age <- simr::powerSim(lmer.full.uk, test = fixed("CDI.z_age_months", method = "z"), seed = 2, nsim = 1000, alpha = 0.05) #specify that Gender is the ixed effect that we are looking into

check_pwr_uk_cdi_age

check_pwr_uk_gender <- simr::powerSim(lmer.full.uk, test = fixed("gender", method = "z"), seed = 3, nsim = 1000, alpha = 0.05) #specify that Gender is the ixed effect that we are looking into

check_pwr_uk_gender
```

### Combined Sample

For this combined analysis, we first need to restrain the age range for the NAE sample (previously ±2 months, now ±1.5 months).

```{r NAE_agerange}
# Create dataset with British and NAE only
data.uk_nae <- data.total %>%
  subset(language_zone %in% c("British", "NAE")) %>%
  mutate(CDI.agemin = ifelse(language_zone == "NAE",
                             CDI.agemin + round(.5*365.25/12),
                             CDI.agemin),
         CDI.agemax = ifelse(language_zone == "NAE",
                             CDI.agemax - round(.5*365.25/12),
                             CDI.agemax)) %>%
  subset(!(CDI.agedays < CDI.agemin | CDI.agedays > CDI.agemax)) %>%
  droplevels()
# Create contrasts for analysis
contrasts(data.uk_nae$gender) <- contr.sum(2)
contrasts(data.uk_nae$method) <- contr.sum(3)
contrasts(data.uk_nae$language_zone) <- contr.sum(2)
```

We can then run the planned combined analysis adding the main effect and interactions of `language_zone`.

```{r uk_nae_lmer}
lmer.full.uk_nae <- lmer(CDI.prop ~ CDI.z_age_months + language_zone + gender +
                           z_age_months + method + IDS_pref + IDS_pref:language_zone +
                           IDS_pref:method + IDS_pref:CDI.z_age_months + IDS_pref:z_age_months +
                           (1 + CDI.z_age_months | labid) + (1 | subid_unique),
                         data = data.uk_nae)

summary(lmer.full.uk_nae)

combined.full.uk_nae_pvalue <- anova(lmer.full.uk_nae) %>% 
  as_tibble(rownames = "Parameter") #this gives us the Type III p values

#==========
# Sequentially removed random effects:
# IDS_pref:z_age_months
# IDS_pref:CDI.z_age_months
# IDS_pref:method
# IDS_pref:language_zone
# IDS_pref
# method
# z_age_months
# gender
# language_zone
#==========

```
#### (Optional) Checking mixed-model assumptions
1) Linearlity
2) Normality of the residuals
3) Homoscedasticity of residuals
4) No autocorrelation
5) No multicollinearity
```{r}
#First, check linearity. The plot looks linear
data.uk_nae$resid <- residuals(lmer.full.uk_nae)
 
plot(data.uk_nae$resid, data.uk_nae$CDI.prop)

#Second, check normality
plot_model(lmer.full.uk_nae, type = 'diag') ## we do have right-skewed normality of residuals

#Third, check autocorrelation
re_run_lme.full.uk_nae <- nlme::lme(CDI.prop ~ CDI.z_age_months + language_zone + gender +
                           z_age_months + method + IDS_pref + IDS_pref:language_zone +
                           IDS_pref:method + IDS_pref:CDI.z_age_months + IDS_pref:z_age_months,
                           random = ~CDI.z_age_months  | labid,
                        method = "REML",
                      data = data.uk_nae, na.action = na.exclude)

plot(nlme::ACF(re_run_lme.full.uk_nae, resType = "normalized")) #there is no sign for autocorrelation
      
#Lastly, check multi-collinearity
car::vif(lmer.full.uk_nae) #no problem for multicollinearlity
```


We then compute $p$-values, but leave out power estimates for those $p$-values as above. Again, we have a lot of singular fit issues for the power checks and decided not to remove parameters based on posthoc power analysis. 

```{r uk_nae_statistics, eval = FALSE}
check_pwr_combined_cdi_age <- simr::powerSim(lmer.full.uk_nae, test = fixed("CDI.z_age_months", method = "z"), seed = 2, nsim = 1000, alpha = 0.05) #specify that Gender is the ixed effect that we are looking into

check_pwr_combined_cdi_age

check_pwr_combined_lang_zone <- simr::powerSim(lmer.full.uk_nae, test = fixed("language_zone", method = "z"), seed = 3, nsim = 1000, alpha = 0.05) #specify that Gender is the ixed effect that we are looking into

check_pwr_combined_lang_zone

check_pwr_combined_gender <- simr::powerSim(lmer.full.uk_nae, test = fixed("gender", method = "z"), seed = 3, nsim = 1000, alpha = 0.05) #specify that Gender is the ixed effect that we are looking into

check_pwr_combined_gender
```


# Bayesian Statistics

Given that the `frequentist#' tests above did not discover any expected significant effects, we now run a Bayesian analysis, specifically Bayes factors (or $b$-values) from model comparisons, to determine whether our data provide evidence for a true null effect or merely fail to provide evidence either way.

We first run models on the separate UK and NAE samples.

First of all, we need to set up the priors for the Bayeisan analysis. In the RR, we said that we would use a Cauchy distribution for priors over fixed effects with normally distributed priors over random effect parameters."

## Set up priors

Set up a really weak and not too informative prior
```{r weak prior}
prior <-c(set_prior("normal(10, 1)", class = "Intercept"), #we can't have negative CDI
              set_prior("cauchy(0, 1)", class = "b"), #all IVs have the same prior..
              set_prior("normal(0, 1)", class = "sd")) 

prior_combined_mod <-c(set_prior("normal(10, 1)", class = "Intercept"),
              set_prior("cauchy(0, 1)", class = "b"), #all IVs have the same prior..
              set_prior("normal(0, 1)", class = "sd"),
              set_prior("lkj(2)", class = "L")) #adding this as there is a random slope and we need a correlation matrix prior between slope and intercept
```

Set up another prior that mainly changed the meaan for the fixed effect and see if there is big difference? 
Note: numbers changed but not to the extent that change BF conclusion.
```{r weak prior test sensitivity}
prior_big <-c(set_prior("normal(10, 1)", class = "Intercept"), #we can't have negative CDI
              set_prior("cauchy(10, 5)", class = "b"), #all IVs have the same prior..
              set_prior("normal(0, 1)", class = "sd")) 

prior_combined_big <-c(set_prior("normal(10, 1)", class = "Intercept"),
              set_prior("cauchy(10, 5)", class = "b"), #all IVs have the same prior..
              set_prior("normal(0, 1)", class = "sd"),
              set_prior("lkj(2)", class = "L")) #adding this as there is a random slope and we need a correlation matrix prior between slope and intercept
```


We also try more informative prior.

## NAE info prior
```{r informative NAE prior}
prior_nae_1 <-c(set_prior("normal(50, 10)", class = "Intercept"), #note that nae's DV is percentile score, but I will assume at 18 months, the average is at 50 percentile
              set_prior("cauchy(5, 1)", class = "b", coef = "CDI.z_age_months"), #assume increase in 1 month may lead to 5 % percentile increase?
              set_prior("cauchy(-15, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males
              set_prior("cauchy(0, 1)", class = "b", coef = "IDS_pref"), #weak prior as don't know the effect of IDS preference to CDI score
              set_prior("cauchy(0, 1)", class = "b", coef = "method1"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 1)", class = "b", coef = "method2"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores
              set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
               set_prior("normal(0, 1)", class = "sd")) 

prior_nae_null_1 <-c(set_prior("normal(50, 10)", class = "Intercept"), #note that nae's DV is percentile score, but I will assume at 18 months, the average is at 50 percentile
              set_prior("cauchy(5, 1)", class = "b", coef = "CDI.z_age_months"), #assume increase in 1 month may lead to 5 % percentile increase?
              set_prior("cauchy(-15, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males
              set_prior("cauchy(0, 1)", class = "b", coef = "method1"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 1)", class = "b", coef = "method2"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores
              set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
               set_prior("normal(0, 1)", class = "sd")) 

prior_nae_2 <-c(set_prior("normal(50, 10)", class = "Intercept"), #note that nae's DV is percentile score, but I will assume at 18 months, the average is at 50 percentile
              set_prior("cauchy(5, 1)", class = "b", coef = "CDI.z_age_months"), #assume increase in 1 month may lead to 5 % percentile increase?
              set_prior("cauchy(-15, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males
              set_prior("cauchy(0, 1)", class = "b", coef = "IDS_pref"), #weak prior as don't know the effect of IDS preference to CDI score
              set_prior("cauchy(0, 1)", class = "b", coef = "method1"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 1)", class = "b", coef = "method2"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores
              #note all interaction priors are weak as we don't really have much info to form a prior
              set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
              set_prior("cauchy(0, 1)", class = "b", coef = "CDI.z_age_months:IDS_pref"),
              set_prior("cauchy(0, 1)", class = "b", coef = "method1:IDS_pref"),
              set_prior("cauchy(0, 1)", class = "b", coef = "method2:IDS_pref"),
              set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months:IDS_pref"),
              set_prior("normal(0, 1)", class = "sd")) 

prior_nae_null_2 <-c(set_prior("normal(50, 10)", class = "Intercept"), #note that nae's DV is percentile score, but I will assume at 18 months, the average is at 50 percentile
              set_prior("cauchy(5, 1)", class = "b", coef = "CDI.z_age_months"), #assume increase in 1 month may lead to 5 % percentile increase?
              set_prior("cauchy(-15, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males
              set_prior("cauchy(0, 1)", class = "b", coef = "IDS_pref"), #weak prior as don't know the effect of IDS preference to CDI score
              set_prior("cauchy(0, 1)", class = "b", coef = "method1"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 1)", class = "b", coef = "method2"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores
               #note all interaction priors are weak as we don't really have much info to form a prior
              set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
              set_prior("cauchy(0, 1)", class = "b", coef = "CDI.z_age_months:IDS_pref"),
              set_prior("cauchy(0, 1)", class = "b", coef = "method1:IDS_pref"),
              set_prior("cauchy(0, 1)", class = "b", coef = "method2:IDS_pref"),
              set_prior("normal(0, 1)", class = "sd")) 
```


## UK info prior
```{r}
prior_uk_1 <-c(set_prior("normal(40, 20)", class = "Intercept"), #note that UK's DV is raw score, based on wordbank (oxford CDI), the average (50 percentile) is around 41.6 
              set_prior("cauchy(15, 1)", class = "b", coef = "CDI.z_age_months"), #assume increase in 1 month may lead to 15 words increase (proxy from wordbank)
              set_prior("cauchy(-10, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males
              set_prior("cauchy(0, 1)", class = "b", coef = "IDS_pref"), #weak prior as don't know the effect of IDS preference to CDI score
              set_prior("cauchy(0, 1)", class = "b", coef = "method1"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
              set_prior("normal(0, 1)", class = "sd")) 

prior_uk_null_1 <-c(set_prior("normal(40, 20)", class = "Intercept"), #note that UK's DV is raw score, based on wordbank (oxford CDI), the average (50 percentile) is around 41.6 
              set_prior("cauchy(15, 1)", class = "b", coef = "CDI.z_age_months"), #assume increase in 1 month may lead to 15 words increase (proxy from wordbank)
              set_prior("cauchy(-10, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males
              set_prior("cauchy(0, 1)", class = "b", coef = "method1"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
              set_prior("normal(0, 1)", class = "sd")) 

prior_uk_2 <-c(set_prior("normal(40, 20)", class = "Intercept"), #note that UK's DV is raw score, based on wordbank (oxford CDI), the average (50 percentile) is around 41.6 
              set_prior("cauchy(15, 1)", class = "b", coef = "CDI.z_age_months"), #assume increase in 1 month may lead to 15 words increase (proxy from wordbank)
              set_prior("cauchy(-10, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males
              set_prior("cauchy(0, 1)", class = "b", coef = "IDS_pref"), #weak prior as don't know the effect of IDS preference to CDI score
              set_prior("cauchy(0, 1)", class = "b", coef = "method1"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
               #note all interaction priors are weak as we don't really have much info to form a prior
              set_prior("cauchy(0, 1)", class = "b", coef = "CDI.z_age_months:IDS_pref"),
              set_prior("cauchy(0, 1)", class = "b", coef = "method1:IDS_pref"),
              set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months:IDS_pref"),
              set_prior("normal(0, 1)", class = "sd")) 

prior_uk_null_2 <-c(set_prior("normal(40, 20)", class = "Intercept"), #note that UK's DV is raw score, based on wordbank (oxford CDI), the average (50 percentile) is around 41.6 
              set_prior("cauchy(15, 1)", class = "b", coef = "CDI.z_age_months"), #assume increase in 1 month may lead to 15 words increase (proxy from wordbank)
              set_prior("cauchy(-10, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males
              set_prior("cauchy(0, 1)", class = "b", coef = "IDS_pref"), #weak prior as don't know the effect of IDS preference to CDI score
              set_prior("cauchy(0, 1)", class = "b", coef = "method1"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
               #note all interaction priors are weak as we don't really have much info to form a prior
              set_prior("cauchy(0, 1)", class = "b", coef = "CDI.z_age_months:IDS_pref"),
              set_prior("cauchy(0, 1)", class = "b", coef = "method1:IDS_pref"),
              set_prior("normal(0, 1)", class = "sd")) 
```

## Combined info prior
```{r}

# Note that NAE CDI WS total score is 680 and UK CDI total score is 416. 

# NAE average scores for 50th percentile is 76.3, so it is 76.3/680 ~ 11% and UK CDI average scores for 50th percentile is 41.6, so it is 41.6/416 ~ 10%

prior_combined_full_info <-c(set_prior("normal(0.1, 0.05)", class = "Intercept"), # given the above info, we expect on average, proportion score is 10% 
              set_prior("cauchy(0.03, 0.01)", class = "b", coef = "CDI.z_age_months"), #assume increase in 1 month may lead to 15 words increase in UK cdi and lead to 28 words increase in NAE (proxy from wordbank). In general, 15/416 = 3% and 28/680 = 4% 
              set_prior("cauchy(-0.03, 0.01)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males, 11/416 ~ 2.6% and Wordbank NAE suggested females knows approximately 27 words than males, 27/680 ~ 3.9%
               set_prior("cauchy(0, 0.01)", class = "b", coef = "language_zone1"), #weak prior as don't know the effect of language zone to CDI score 
              set_prior("cauchy(0, 0.01)", class = "b", coef = "IDS_pref"), #weak prior as don't know the effect of IDS preference to CDI score
              set_prior("cauchy(0, 0.01)", class = "b", coef = "method1"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 0.01)", class = "b", coef = "method2"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 0.01)", class = "b", coef = "z_age_months"),
               #note all interaction priors are weak as we don't really have much info to form a prior
              set_prior("cauchy(0, 0.01)", class = "b", coef = "CDI.z_age_months:IDS_pref"),
              set_prior("cauchy(0, 0.01)", class = "b", coef = "language_zone1:IDS_pref"),
              set_prior("cauchy(0, 0.01)", class = "b", coef = "method1:IDS_pref"),
              set_prior("cauchy(0, 0.01)", class = "b", coef = "method2:IDS_pref"),
              set_prior("cauchy(0, 0.01)", class = "b", coef = "z_age_months:IDS_pref"),
              set_prior("normal(0, 0.01)", class = "sd"),
              set_prior("lkj(2)", class = "L")) #adding this as there is a random slope and we need a correlation matrix prior between slope and intercept

prior_combined_null_info <-c(set_prior("normal(0.1, 0.05)", class = "Intercept"), # given the above info, we expect on average, proportion score is 10% 
              set_prior("cauchy(0.03, 0.01)", class = "b", coef = "CDI.z_age_months"), #assume increase in 1 month may lead to 15 words increase in UK cdi and lead to 28 words increase in NAE (proxy from wordbank). In general, 15/416 = 3% and 28/680 = 4% 
              set_prior("cauchy(-0.03, 0.01)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males, 11/416 ~ 2.6% and Wordbank NAE suggested females knows approximately 27 words than males, 27/680 ~ 3.9%
               set_prior("cauchy(0, 0.01)", class = "b", coef = "language_zone1"), #weak prior as don't know the effect of language zone to CDI score 
              set_prior("cauchy(0, 0.01)", class = "b", coef = "IDS_pref"), #weak prior as don't know the effect of IDS preference to CDI score
              set_prior("cauchy(0, 0.01)", class = "b", coef = "method1"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 0.01)", class = "b", coef = "method2"), #weak prior as methods related to the MB1 study shouldn't influence CDI scores 
              set_prior("cauchy(0, 0.01)", class = "b", coef = "z_age_months"),
               #note all interaction priors are weak as we don't really have much info to form a prior
              set_prior("cauchy(0, 0.01)", class = "b", coef = "CDI.z_age_months:IDS_pref"),
              set_prior("cauchy(0, 0.01)", class = "b", coef = "method1:IDS_pref"),
              set_prior("cauchy(0, 0.01)", class = "b", coef = "method2:IDS_pref"),
              set_prior("cauchy(0, 0.01)", class = "b", coef = "z_age_months:IDS_pref"),
              set_prior("normal(0, 0.01)", class = "sd"),
              set_prior("lkj(2)", class = "L")) #adding this as there is a random slope and we need a correlation matrix prior between slope and intercept
```


In the following analysis, we follow what we have proposed in the RR about calculating the Bayes factor for the three variables of interest:
1) The main effect IDS preference on CDI scores
2) The interaction effect between IDS preference and IDS test age on CDI scores
3) The interaction effect between language_zone (i.e., dialect in RR) and IDS preference on CDI scores

For the first two variables, we run it on the NAE and UK model separately. For the third variable, we run in on the combined UK & NAE model. 

To calcualte Bayes factor for the main effect IDS preference on CDI scores, we need to two models: a full model that contains the main effect of variables of interest, and the other null model that does not contain the main effect of interest. As all the interaction terms in the models contains IDS preference, in the following, we only include main effects in the full and null models for comparisons.

## NAE model (main effect of IDS preference)
```{r}
Sys.setenv(R_FUTURE_RNG_ONMISUSE = "ignore") #this line of code is needed to get rid of warning message due to the "future" package. https://github.com/satijalab/seurat/issues/3622 

lmer.bf_full_1.nae <- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
                        z_age_months + method + IDS_pref +
                        (1 | labid) + (1 | subid_unique),
                      data = data.nae, 
                      family = gaussian,
                      prior = prior_nae_1,
                      iter = 4000, 
                      warmup = 2000,
                      chains = 4,
                      save_pars = save_pars(all = T),
                      control = list(adapt_delta = .95), 
                      seed=456)

summary(lmer.bf_full_1.nae)
                      
lmer.bf_null_1.nae <- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
                        z_age_months + method + 
                        (1 | labid) + (1 | subid_unique),
                      data = data.nae, 
                      family = gaussian,
                      prior = prior_nae_null_1,
                      iter = 4000, 
                      warmup = 2000,
                      chains = 4,
                      save_pars = save_pars(all = T),
                      control = list(adapt_delta = .95), 
                      seed =123)

summary(lmer.bf_null_1.nae)
```

Calculate the marginal log likelihoods for each hypothese using Bridge sampling
```{r}

marloglik_full_1.nae <- bridge_sampler(lmer.bf_full_1.nae, silent = T)
marloglik_null_1.nae <- bridge_sampler(lmer.bf_null_1.nae, silent = T)

```

Bayes factor
```{r}

BF_full_1.nae <- bayes_factor(marloglik_full_1.nae, marloglik_null_1.nae)
BF_full_1.nae

```

## NAE model (interaction between IDS test age and IDS preference)
```{r}

lmer.bf_full_2.nae <- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
                        z_age_months + method + IDS_pref +
                        IDS_pref:method + IDS_pref:CDI.z_age_months + IDS_pref:z_age_months +
                        (1 | labid) + (1 | subid_unique),
                      data = data.nae, 
                      family = gaussian,
                      prior = prior_nae_2,
                      iter = 4000, 
                      warmup = 2000,
                      chains = 4,
                      save_pars = save_pars(all = T),
                      control = list(adapt_delta = .95), 
                      seed=890)

summary(lmer.bf_full_2.nae)
                      

lmer.bf_null_2.nae <- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
                        z_age_months + method + IDS_pref +
                        IDS_pref:method + IDS_pref:CDI.z_age_months + 
                        (1 | labid) + (1 | subid_unique),
                      data = data.nae, 
                      family = gaussian,
                      prior = prior_nae_null_2,
                      iter = 4000, 
                      warmup = 2000,
                      chains = 4,
                      save_pars = save_pars(all = T),
                      control = list(adapt_delta = .95), 
                      seed =102)

summary(lmer.bf_null_2.nae)
```

Calculate the marginal log likelihoods for each hypothese using Bridge sampling
```{r}

marloglik_full_2.nae <- bridge_sampler(lmer.bf_full_2.nae, silent = T)
marloglik_null_2.nae <- bridge_sampler(lmer.bf_null_2.nae, silent = T)

```

Bayes factor
```{r}

BF_full_2.nae <- bayes_factor(marloglik_full_2.nae, marloglik_null_2.nae)
BF_full_2.nae

```

## UK model (main effect of IDS preference)
```{r}

lmer.bf_full_1.uk <- brm(vocab_nwords ~ CDI.z_age_months + gender +
                       z_age_months + method + IDS_pref +
                      (1 | labid) + (1 | subid_unique),
                     data = data.uk, 
                      family = gaussian,
                      prior = prior_uk_1,
                      iter = 4000, 
                      warmup = 2000,
                      chains = 4,
                      save_pars = save_pars(all = T),
                      control = list(adapt_delta = .95), 
                      seed=213)

summary(lmer.bf_full_1.uk)
                      

lmer.bf_null_1.uk <- brm(vocab_nwords ~ CDI.z_age_months + gender +
                       z_age_months + method + 
                      (1 | labid) + (1 | subid_unique),
                      data = data.uk, 
                      family = gaussian,
                      prior = prior_uk_null_1,
                      iter = 4000, 
                      warmup = 2000,
                      chains = 4,
                      save_pars = save_pars(all = T),
                      control = list(adapt_delta = .95), 
                      seed =312)

summary(lmer.bf_null_1.uk)
```

Calculate the marginal log likelihoods for each hypothese using Bridge sampling
```{r}
marloglik_full_1.uk <- bridge_sampler(lmer.bf_full_1.uk, silent = T)
marloglik_null_1.uk <- bridge_sampler(lmer.bf_null_1.uk, silent = T)
```

Bayes factor
```{r}
BF_full_1.uk <- bayes_factor(marloglik_full_1.uk, marloglik_null_1.uk)
BF_full_1.uk
```

## UK model (interaction between IDS test age and IDS preference)
```{r}


lmer.bf_full_2.uk <- brm(vocab_nwords ~ CDI.z_age_months + gender +
                       z_age_months + method + IDS_pref +
                       IDS_pref:method + IDS_pref:CDI.z_age_months + IDS_pref:z_age_months +
                      (1 | labid) + (1 | subid_unique),
                     data = data.uk, 
                      family = gaussian,
                      prior = prior_uk_2,
                      iter = 4000, 
                      warmup = 2000,
                      chains = 4,
                      save_pars = save_pars(all = T),
                      control = list(adapt_delta = .95), 
                      seed=546)

summary(lmer.bf_full_2.uk)
                      

lmer.bf_null_2.uk <- brm(vocab_nwords ~ CDI.z_age_months + gender +
                       z_age_months + method + IDS_pref +
                       IDS_pref:method + IDS_pref:CDI.z_age_months + 
                      (1 | labid) + (1 | subid_unique),
                      data = data.uk, 
                      family = gaussian,
                      prior = prior_uk_null_2,
                      iter = 4000, 
                      warmup = 2000,
                      chains = 4,
                      save_pars = save_pars(all = T),
                      control = list(adapt_delta = .95), 
                      seed =689)

summary(lmer.bf_null_2.uk)
```

Calculate the marginal log likelihoods for each hypothese using Bridge sampling
```{r}
marloglik_full_2.uk <- bridge_sampler(lmer.bf_full_2.uk, silent = T)
marloglik_null_2.uk <- bridge_sampler(lmer.bf_null_2.uk, silent = T)
```

Bayes factor
```{r}
BF_full_2.uk <- bayes_factor(marloglik_full_2.uk, marloglik_null_2.uk)
BF_full_2.uk
```

## Combined model (interaction between lanaguage zone and IDS preference)
```{r Bayesian_combined Bayes Factor analysis}


lmer.bf_full_3.combined <- brm(CDI.prop ~ CDI.z_age_months + language_zone + gender +
                           z_age_months + method + IDS_pref + IDS_pref:language_zone +
                           IDS_pref:method + IDS_pref:CDI.z_age_months + IDS_pref:z_age_months +
                           (1 + CDI.z_age_months | labid) + (1 | subid_unique),
                         data = data.uk_nae,
                        family = gaussian,
                      prior = prior_combined_full_info,
                      iter = 4000, 
                      warmup = 2000,
                      chains = 4,
                      save_pars = save_pars(all = T),
                      control = list(adapt_delta = .9999), 
                      seed=100)

summary(lmer.bf_full_3.combined)
                      

lmer.bf_null_3.combined <- brm(CDI.prop ~ CDI.z_age_months + language_zone + gender +
                           z_age_months + method + IDS_pref + 
                           IDS_pref:method + IDS_pref:CDI.z_age_months + IDS_pref:z_age_months +
                           (1 + CDI.z_age_months | labid) + (1 | subid_unique),
                         data = data.uk_nae,
                      family = gaussian,
                      prior = prior_combined_null_info,
                      iter = 4000, 
                      warmup = 2000,
                      chains = 4,
                      save_pars = save_pars(all = T),
                      control = list(adapt_delta = .9999), 
                      seed =200)

summary(lmer.bf_null_3.combined)
```

Calculate the marginal log likelihoods for each hypothese using Bridge sampling
```{r}
marloglik_full_3.combined <- bridge_sampler(lmer.bf_full_3.combined, silent = T)
marloglik_null_3.combined <- bridge_sampler(lmer.bf_null_3.combined, silent = T)
```

Bayes factor
```{r}
BF_full_3.combined <- bayes_factor(marloglik_full_3.combined, marloglik_null_3.combined)
BF_full_3.combined
```
```
