==> 03a_confirmatory_analysis.Rmd <==
---
title: "MB1 CDI Follow-up Confirmatory Analyses: Freqentist"
author: "The ManyBabies Analysis Team"
date: '`r format(Sys.time(), "%a %b %d %X %Y")`'
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: yes
---

```{r session_infos}
# Provides a session log for the packages used in this .rmd
library(sessioninfo)
session_info(pkgs = "!attached", to_file = "03a_session_log.txt")
```

# Introduction

In this script, we go through all the pre-registered proposed analyses. As a reminder, the main research questions where as follows:

1. To what extent does infants' preference for IDS as measured in a laboratory setting predict their vocabulary at 18 and 24 months?
2. Does the relation between IDS preference and vocabulary size change over development?
3. Are there systematic differences in the strength of this relationship across the language communities in our sample?

Here we present the main "sample theory based" analyses (also known as frequentist), separately on the North American and UK samples in parallel to answer our first two research questions, then together to answer our third research question. In the next section (03b) provide additional Bayesian statistics where a null effect was found, as specified in the pre-registration.

```{r setup, message=FALSE, warning=FALSE}
# Library imports, general settings ==============
library(tidyverse)
library(egg)
library(knitr)
library(lme4)
library(lmerTest)
library(simr)
library(modelsummary)  #<= creates and exports the  tables for the models as word doc
 
# As in our discussion with Mike, we will use lmerTest for calculating p values
library(lattice)
library(effects)
library(sjPlot)
library(robustlmm)
library(car)
library(papaja)
theme_apa <- papaja::theme_apa

# Load model comparison functions
source("helper/lrtests.R")

# Deal with package priority issues
select <- dplyr::select

theme_set(theme_bw(base_size = 10))
options("future" = T)
# knitr::opts_chunk$set(cache = TRUE)

print(sessionInfo()) # listing all info about R and packages info

# Read data ======================================
col_types <- cols(
  labid = col_factor(),
  subid = col_factor(),
  subid_unique = col_factor(),
  CDI.form = col_factor(),
  CDI.nwords = col_integer(),
  CDI.prop = col_number(),
  CDI.agerange = col_factor(),
  CDI.agedays = col_integer(),
  CDI.agemin = col_integer(),
  CDI.agemax = col_integer(),
  vocab_nwords = col_integer(),
  standardized.score.CDI = col_character(),
  standardized.score.CDI.num = col_number(),
  z.IDS_pref = col_number(),
  language = col_factor(),
  language_zone = col_factor(),
  CDI.error = col_logical(),
  Notes = col_character(),
  trial_order = col_factor(),
  method = col_factor(),
  age_days = col_integer(),
  age_mo = col_number(),
  age_group = col_factor(),
  nae = col_logical(),
  gender = col_factor(),
  second_session = col_logical()
)
data.total <- read_csv("data/02b_processed.csv", col_types = col_types)
```

Before moving on with the analysis, we have to ready the data by (a) checking for colinearity between `z_age_months` and `CDI.z_age_months` and correcting this if necessary, and (b) setting up the contrasts described in our data analysis.

## Colinearity check

First, we run a Kappa test on the possibility of colinearity between `z_age_months` and `CDI.z_age_months`.

```{r colinearity_check}
# Run kappa test
k.age_months <- model.matrix(~ z_age_months + CDI.z_age_months, data = data.total) %>%
  kappa(exact = T)
```

With a value of `r k.age_months`, we do not have a colinearity issue and can proceed with the data analysis as planned (The criteria of indicating colinearity is that kappa > 10). 

## Contrast Setups

We need `gender` as an effect-coded factor, and `method` as a deviation-coded factor. This is achieved in R by using the `contr.sum()` function with the number of levels for each factor. Notably, when subsetting the UK sample, only two levels of `method` out of the three in total were left. 

```{r contrasts}
# Set contrasts on the total dataset =============
contrasts(data.total$gender) <- contr.sum(2)
contrasts(data.total$method) <- contr.sum(3)

# Create sub-datasets, with contrasts ============
## NAE
data.nae <- data.total %>%
  subset(language_zone == "NAE") %>%
  droplevels()
contrasts(data.nae$gender) <- contr.sum(2)
contrasts(data.nae$method) <- contr.sum(3)

## UK (combined-age and separate 18/24 months)

data.uk <- data.total %>%
  subset(language_zone == "British") %>%
  droplevels()
contrasts(data.uk$gender) <- contr.sum(2)
contrasts(data.uk$method) <- contr.sum(2) # note that UK sample has only 2 levels, so sum of zero contrasts set to 2 levels


data.uk.18 <- data.total %>%
  subset(language_zone == "British" & CDI.agerange ==
    "18") %>%
  droplevels()
contrasts(data.uk.18$gender) <- contr.sum(2)
contrasts(data.uk.18$method) <- contr.sum(2) # note that UK sample has only 2 levels, so sum of zero contrasts set to 2 levels

data.uk.24 <- data.total %>%
  subset(language_zone == "British" & CDI.agerange ==
    "24") %>%
  droplevels()
contrasts(data.uk.24$gender) <- contr.sum(2)
contrasts(data.uk.24$method) <- contr.sum(2) # note that UK sample has only 2 levels, so sum of zero contrasts set to 2 levels



## Other
data.other <- data.total %>%
  subset(language_zone == "Other") %>%
  droplevels()
contrasts(data.other$gender) <- contr.sum(2)
contrasts(data.other$method) <- contr.sum(3)
```

# Descriptive Statistics

We first assess the amount of data we have overall per condition and their shape overall.

```{r desc_total}
data.total %>%
  group_by(language_zone, CDI.agerange, method, gender) %>%
  summarise(N = n(), age = mean(CDI.agedays), sd = sd(CDI.agedays)) %>%
  kable()
```

Total number of children by age group

```{r}
data.total %>%
  group_by(CDI.agerange) %>%
  summarise(N = n()) %>%
  kable()
```

Total number of CDIs (not unique children)

```{r}
data.total %>%
  summarise(N = n()) %>%
  kable()
```
Total number of CDIs (not unique children) by language zone

```{r}
data.total %>%
  group_by(language_zone) %>%
  summarise(N = n()) %>%
  kable()
```

Total number of unique children

```{r}
data.total %>%
  distinct(subid_unique) %>%
  summarise(N = n()) %>%
  kable()
```

Total number of unique children per language zone

```{r}
data.total %>%
  group_by(language_zone) %>%
  distinct(subid_unique) %>%
  summarise(N = n()) %>%
  kable()
```
Total number of children per gender and age

```{r}
data.total %>%
  group_by(gender,CDI.agerange) %>%
  summarise(N = n()) %>%
  kable()
```
Total number of children per gender and age

```{r}
data.total %>%
  group_by(gender) %>%
  summarise(N = n()) %>%
  kable()
```
More detailed information about Descriptive Statistics

```{r}
# number of lab
data.total %>%
  select(labid, language_zone) %>%
  unique() %>%
  group_by(language_zone) %>%
  count()

data.total %>%
  group_by(language_zone, CDI.agerange) %>%
  summarize(N = n())

# age range in each age group and language zone
data.total %>%
  select(subid, language_zone, CDI.agedays, CDI.agerange) %>%
  unique() %>%
  group_by(language_zone, CDI.agerange) %>%
  summarize(
    age_min = (min(CDI.agedays) / 365.25 * 12),
    age_max = (max(CDI.agedays) / 365.25 * 12)
  )
```

We then assess the data per lab in terms of sample size and CDI score (vocabulary size, for consistency between language zones).


```{r desc_by_lab}
by_lab <- data.total %>%
  group_by(labid, language_zone, CDI.agerange) %>%
  mutate(tested = n_distinct(subid_unique)) %>%
  select(labid, language_zone, CDI.agerange, tested, vocab_nwords) %>%
  nest(scores = vocab_nwords) %>%
  mutate(
    model = map(scores, ~ lm(vocab_nwords ~ 1, data = .x)),
    ci = map(model, confint)
  ) %>%
  transmute(
    tested = tested,
    mean = map_dbl(model, ~ coefficients(.x)[[1]]),
    ci_lower = map_dbl(ci, 1),
    ci_upper = map_dbl(ci, 2)
  ) %>%
  arrange(language_zone) %>%
  rownames_to_column() %>%
  ungroup()
```

## Visualization by Lab (Mean and confidence intervals)

```{r desc_by_lab_vis}
# created a new column with the labs IDs as character so it can be sorted in alphabetical order
by_lab$labid_car <- as.character(by_lab$labid)

# relabel the CDI age factor column
levels(by_lab$CDI.agerange) <- c("18-Month Olds", "24-Month Olds")

# Making sure the factor columns have levels in the order I would like to graph them
by_lab$language_zone <- factor(by_lab$language_zone, levels = c("Other", "NAE", "British"))

# sorted the idcolum in the way I would it to show in the ggplot
labid_ord <- by_lab %>%
  dplyr::arrange(language_zone, desc(labid_car)) %>%
  ungroup() %>%
  filter(CDI.agerange == "18-Month Olds") %>%
  select(labid)

# Making sure the factor columns have levels in the order I would like to graph them
by_lab$labid <- factor(by_lab$labid, levels = labid_ord$labid)

## graph by language zone and lab Id in asc order
by_lab %>%
  ggplot(
    .,
    aes(
      x = labid,
      y = mean, colour = language_zone,
      ymin = ci_lower, ymax = ci_upper
    )
  ) +
  geom_linerange() +
  geom_point(aes(size = tested)) +
  coord_flip(ylim = c(0, 500)) +
  xlab("Laboratory ID") +
  ylab("Average Vocabulary Size") +
  scale_colour_brewer(
    palette = "Dark2", name = "Language\nZone",
    breaks = c("British", "NAE", "Other")
  ) +
  scale_size_continuous(name = "Sample\nSize") +
  facet_wrap(vars(CDI.agerange)) +
  theme(  text = element_text(size = 20)) 

ggsave("plots/vocab-size_by-lab.png", width = 12, height = 8)
```


# Sample Theory Based Statistics

## Simple Correlation

First, we want to assess quickly if there is a direct correlation between IDS preference and CDI score, computing a Pearson#'s product-moment correlation. We use standardized CDI scores for the North American sample, and raw scores for the British sample. Since CDI grows with age, we run the British sample separately for 18 and 24 months.

```{r simple_correlation}
# Statistics =====================================
## North American Sample
test.pearson.nae <- cor.test(data.nae$z.IDS_pref,
  data.nae$z_standardized_CDI,
  alternative = "two.sided", method = "pearson"
)

test.pearson.nae

## UK Sample
test.pearson.uk.18 <- cor.test(data.uk.18$z.IDS_pref,
  data.uk.18$z_vocab_nwords,
  alternative = "two.sided", method = "pearson"
)

test.pearson.uk.18

test.pearson.uk.24 <- cor.test(data.uk.24$z.IDS_pref,
  data.uk.24$z_vocab_nwords,
  alternative = "two.sided", method = "pearson"
)

test.pearson.uk.24
```

### Plots for correlation

```{r}
## North American Sample
### Get correlation value for annotation
cor_text <- "paste(italic(R)^2, \" =\")"
cor_value <- round(test.pearson.nae$estimate, 3)


### Build plot

xrng <- range(data.nae$IDS_pref)

plot.pearson.nae <- data.nae %>%
  ggplot(aes(
    x = IDS_pref,
    y = standardized.score.CDI.num
  )) +
  xlab("Standardized IDS Preference") +
  ylab("Standardized CDI Score") +
  labs(title="NAE") +
  geom_point(colour = "#D95F02") +
  geom_smooth(method = lm, color = "#D95F02") +
  annotate("text",x = xrng[2], y = 100,
   hjust = 1, vjust = 0, parse = T, size = 3,
    label = paste(cor_text, cor_value, sep = "~")
  ) + theme_apa()

## UK Sample
cor_value_18 <- round(test.pearson.uk.18$estimate, 3)
cor_value_24 <- round(test.pearson.uk.24$estimate, 3)

plot.pearson.uk <- data.uk %>%
  ggplot(aes(
    x = IDS_pref,
    y = vocab_nwords,
    colour = CDI.agerange
  )) +
  xlab("Standardized IDS Preference") +
  ylab("Vocabulary Size (Total Words)") +
  labs(title="UK") +
  labs(colour = "Age \nGroup") +
  geom_point() +
  geom_smooth(method = lm ) +
    scale_color_manual(values = c("#D81159","#006BA6")) +
  annotate("text",x = xrng[2], y = 415,
   hjust = 1, vjust = 0, parse = T, size = 3,
    label = paste(cor_text, cor_value_24, sep = "~")) +
  annotate("text",x = xrng[2], y = 1,
   hjust = 1, vjust = 0, parse = T, size = 3,
    label = paste(cor_text, cor_value_18,sep = "~")) + theme_apa() +  ylim(0,420)
  


# Global plot
plot.pearson <- ggarrange(plot.pearson.nae, plot.pearson.uk, ncol = 2)

plot.pearson


ggsave("plots/corr_plot.png", plot.pearson,
  units = "mm", width = 180, height = 100, dpi = 1000
)
```

We see no obvious direct link between IDS prefernce and CDI score here. However, an effect might appear once we take into account various factors that might interact with IDS preference and/or CDI score. We can also first enhance these plots with information about the age group at which infants were tested (18- or 24-month-old) for the NAE sample, using vocabulary size to better compare the NAE and UK samples.


## Mixed-Effects Model by Language Zone

Here, we run a mixed-effects model including only theoretically motivated effects, as described in the pre-registration. We start with the full model bellow, simplifying the random effects structure until it converges.

### NAE full model

```{r full_lmer}
# Run models =====================================
## NAE


lmer.full.nae <- lmer(standardized.score.CDI.num ~ CDI.z_age_months + gender + z_age_months + method + z.IDS_pref + z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months + (1 | labid) + (1 | subid_unique),
data = data.nae
)

summary(lmer.full.nae)

# robust_lmer.full.nae <- robustlmm::rlmer(standardized.score.CDI.num ~ CDI.z_age_months + gender +
#                         z_age_months + method + z.IDS_pref +
#                         z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
#                         (1 | labid),
#                       data = data.nae)
#
#
# summary(robust_lmer.full.nae) #this model is used to see if we can meet some statistical assumption better but we decided to use the original model as the inferential statistical results are consistent

full.nae_pvalue <- anova(lmer.full.nae) %>%
  as_tibble(rownames = "Parameter") # this gives us the Type III p values

# ==========
# Sequentially removed random effects:
# z.IDS_pref:z_age_months
# z.IDS_pref:CDI.z_age_months
# z.IDS_pref:method
# z.IDS_pref
# method
# z_age_months
# gender
# CDI.z_age_months
# ==========
```

#### (Optional) Checking mixed-model assumptions. We will check the following:
1) Linearlity
2) Normality of the residuals
3) Homoscedasticity of residuals
4) No autocorrelation
5) No multicollinearity
```{r}
# First, check linearity
# data.nae$resid <- residuals(lmer.full.nae)
#
# plot(data.nae$resid, data.nae$standardized.score.CDI)

# Second, check normality
plot_model(lmer.full.nae, type = "diag") ## we do have right-skewed normality of residuals

# Third, check autocorrelation
re_run_lme.full.nae <- nlme::lme(standardized.score.CDI.num ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months,
random = ~ 1 | labid,
method = "REML",
data = data.nae, na.action = na.exclude
)

plot(nlme::ACF(re_run_lme.full.nae, resType = "normalized")) # there is no sign for autocorrelation

# Lastly, check multi-collinearity
car::vif(lmer.full.nae) # we do see a multicollineartiy for the IDS preference variable, even though we have centered the IDS preference score. It is probably related to the number the participating labs (as this is the group level that we are controlling) and how we entered interaction between IDS preference and other variables (that lack variability in the current sample). We need to keep IDS preference in the model as exploring the relationship between IDS preference and CDI score is the key research question in the paper.
```
#### Table Summary
```{r}
modelsummary_lmer_full_nae<- list(
  "Full Model" = lmer.full.nae)

msummary(modelsummary_lmer_full_nae, output = "tables/lmer_full_nae.docx",
             stars = TRUE, align = "lc",metrics=c("RMSE","R2","AIC", "BIC","Log.Lik.","F"))
```


### UK full model
```{r}

lmer.full.uk <- lmer(vocab_nwords ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  # (1 | labid) +
  (1 | subid_unique),
data = data.uk
)

summary(lmer.full.uk)

full.uk_pvalue <- anova(lmer.full.uk) %>%
  as_tibble(rownames = "Parameter") # this gives us the Type III p values

# ==========
# Sequentially removed random effects:
# z.IDS_pref:z_age_months
# z.IDS_pref:CDI.z_age_months
# z.IDS_pref:method
# z.IDS_pref
# method
# z_age_months
# gender
# CDI.z_age_months
# (1| labid) please note the variance was very little and reported as zero in the results, we needed to remove this random effect
```
#### (Optional) Checking mixed-model assumptions. We will check the following:
1) Linearlity
2) Normality of the residuals
3) Homoscedasticity of residuals
4) No autocorrelation
5) No multicollinearity

```{r}
# First, check linearity. The plot looks linear
data.uk$resid <- residuals(lmer.full.uk)

plot(data.uk$resid, data.uk$vocab_nwords)

# Second, check normality
plot_model(lmer.full.uk, type = "diag") ## we do have right-skewed normality of residuals

# Third, check autocorrelation
re_run_lme.full.uk <- nlme::lme(vocab_nwords ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months,
random = ~ 1 | labid,
method = "REML",
data = data.nae, na.action = na.exclude
)

plot(nlme::ACF(re_run_lme.full.uk, resType = "normalized")) # there is no sign for autocorrelation

# Lastly, check multi-collinearity
car::vif(lmer.full.uk) # no problem for multicollinearlity
```


We now want to check the statistical power of significant effects, and discard any models with significant effects that do not reach 80% power. This however leads to too many warnings of singularity issues on the model updates inherent to the `simr` power simulations, hence we cannot obtain satisfactory power estimates as pre-registered.

AST: Note that we don't have any IV(s) that turned out to be significant in the Full NAE model. So we won't run the power analysis check. For the UK full model, there are two statistically significant IV: CDI_age and gender. The post hoc power check suggested that we have high power in detecting the effect of CDI_age but not gender. Note that gender has a smaller effect size to begin with, so this may partially explain why we have less power in detecting it in the model. As there can be a number of different factors that determines the posthoc power, we decided not to remove gender in the model based on posthoc power analysis check. 

```{r pwr_lmer, eval = FALSE}

check_pwr_uk_cdi_age <- simr::powerSim(lmer.full.uk, test = fixed("CDI.z_age_months", method = "z"), seed = 2, nsim = 1000, alpha = 0.05) # specify that Gender is the fixed effect that we are looking into

check_pwr_uk_cdi_age

check_pwr_uk_gender <- simr::powerSim(lmer.full.uk, test = fixed("gender", method = "z"), seed = 3, nsim = 1000, alpha = 0.05) # specify that Gender is the fixed effect that we are looking into

check_pwr_uk_gender
```

#### Table Summary
```{r}
modelsummary_lmer_full_uk<- list(
  "Full Model" = lmer.full.uk)

msummary(modelsummary_lmer_full_uk, output = "tables/lmer_full_uk.docx",
             stars = TRUE, align = "lc",metrics=c("RMSE","R2","AIC", "BIC","Log.Lik.","F"))
```

### Combined Sample

For this combined analysis, we first need to restrain the age range for the NAE sample (previously ±2 months, now ±0.5 months). 

```{r NAE_agerange}
# Create dataset with British and NAE only
before_exclusion_participants <- data.total %>%
  filter(language_zone == "NAE" | language_zone == "British") %>%
  distinct(subid_unique) %>%
  count()
before_exclusion_CDIs <- data.total %>%
  filter(language_zone == "NAE" | language_zone == "British") %>%
  count()

data.uk_nae <- data.total %>%
  subset(language_zone %in% c("British", "NAE")) %>%
  mutate(
    CDI.agemin = ifelse(language_zone == "NAE",
      CDI.agemin + round(.5 * 365.25 / 12),
      CDI.agemin
    ),
    CDI.agemax = ifelse(language_zone == "NAE",
      CDI.agemax - round(.5 * 365.25 / 12),
      CDI.agemax
    )
  ) %>%
  subset(!(CDI.agedays < CDI.agemin | CDI.agedays > CDI.agemax)) %>%
  droplevels()
# Create contrasts for analysis
contrasts(data.uk_nae$gender) <- contr.sum(2)
contrasts(data.uk_nae$method) <- contr.sum(3)
contrasts(data.uk_nae$language_zone) <- contr.sum(2)

after_exclusion_participants <- data.uk_nae %>%
  distinct(subid_unique) %>%
  count()
after_exclusion_CDIs <- count(data.uk_nae)
```

We go from `r before_exclusion_participants` to `r after_exclusion_participants` total participants in the combined sample, meaning that `r before_exclusion_participants - after_exclusion_participants` participants were excluded from the North American sample. In total, `r before_exclusion_CDIs - after_exclusion_CDIs` rows of data were removed.

We can then run the planned combined analysis adding the main effect and interactions of `language_zone`.

```{r uk_nae_lmer}
lmer.full.uk_nae <- lmer(CDI.prop ~ CDI.z_age_months + language_zone + gender +
  z_age_months + method + z.IDS_pref + z.IDS_pref:language_zone +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  (1 | labid) + (1 | subid_unique),
data = data.uk_nae
)

summary(lmer.full.uk_nae)

combined.full.uk_nae_pvalue <- anova(lmer.full.uk_nae) %>%
  as_tibble(rownames = "Parameter") # this gives us the Type III p values

# ==========
# Sequentially removed random effects:
# z.IDS_pref:z_age_months
# z.IDS_pref:CDI.z_age_months
# z.IDS_pref:method
# z.IDS_pref:language_zone
# z.IDS_pref
# method
# z_age_months
# gender
# language_zone
# CDI.z_age_months
# ==========
```
#### (Optional) Checking mixed-model assumptions
1) Linearlity
2) Normality of the residuals
3) Homoscedasticity of residuals
4) No autocorrelation
5) No multicollinearity
```{r}
# First, check linearity. The plot looks linear
data.uk_nae$resid <- residuals(lmer.full.uk_nae)

plot(data.uk_nae$resid, data.uk_nae$CDI.prop)

# Second, check normality
plot_model(lmer.full.uk_nae, type = "diag") ## we do have right-skewed normality of residuals

# Third, check autocorrelation
re_run_lme.full.uk_nae <- nlme::lme(CDI.prop ~ CDI.z_age_months + language_zone + gender +
  z_age_months + method + z.IDS_pref + z.IDS_pref:language_zone +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months,
random = ~ 1 | labid,
method = "REML",
data = data.uk_nae, na.action = na.exclude
)

plot(nlme::ACF(re_run_lme.full.uk_nae, resType = "normalized")) # there is no sign for autocorrelation

# Lastly, check multi-collinearity
car::vif(lmer.full.uk_nae) # no problem for multicollinearlity
```


We then compute $p$-values, but leave out power estimates for those $p$-values as above. Again, we have a lot of singular fit issues for the power checks and decided not to remove parameters based on posthoc power analysis. 

```{r uk_nae_statistics, eval = FALSE}
check_pwr_combined_cdi_age <- simr::powerSim(lmer.full.uk_nae, test = fixed("CDI.z_age_months", method = "z"), seed = 2, nsim = 1000, alpha = 0.05) # specify that Gender is the fixed effect that we are looking into

check_pwr_combined_cdi_age

check_pwr_combined_lang_zone <- simr::powerSim(lmer.full.uk_nae, test = fixed("language_zone", method = "z"), seed = 3, nsim = 1000, alpha = 0.05) # specify that Gender is the fixed effect that we are looking into

check_pwr_combined_lang_zone

check_pwr_combined_gender <- simr::powerSim(lmer.full.uk_nae, test = fixed("gender", method = "z"), seed = 3, nsim = 1000, alpha = 0.05) # specify that Gender is the fixed effect that we are looking into

check_pwr_combined_gender
```

#### Table Summary
```{r}
modelsummary_lmer_full_uk_nae<- list(
  "Full Model" = lmer.full.uk_nae)

msummary(modelsummary_lmer_full_uk_nae, output = "tables/lmer_full_uk_nae.docx",
             stars = TRUE, align = "lc",metrics=c("RMSE","R2","AIC", "BIC","Log.Lik.","F"))
```


### Summary table all 3 models
```{r}
modelsummary_all_three_models<- list(
  "NAE Full Model"= lmer.full.nae,
  "UK Full Model" = lmer.full.uk,
  "NAE+UK Full Model" = lmer.full.uk_nae)

msummary(modelsummary_all_three_models, output = "tables/lmer_full_three_models.docx",
             stars = TRUE, align = "lccc",metrics=c("RMSE","R2","AIC", "BIC","Log.Lik.","F"))
```
