==> 03b_confirmatory_analysis.Rmd <==
---
title: "MB1 CDI Follow-up Confirmatory Analyses: Bayesian"
author: "The ManyBabies Analysis Team"
date: '`r format(Sys.time(), "%a %b %d %X %Y")`'
output: 
  html_document:
    toc: true
    toc_float: true
    number_sections: yes
---

# Introduction

In this script, follow up the Frequentist analyses by providing additional Bayesian statistics where a null effect was found, as specified in the pre-registration.

```{r setup, message=FALSE, warning=FALSE}
# Library imports, general settings ==============
library(tidyverse)
library(egg)
library(knitr)

# in the mixed-effects models (noted as a deviation)
library(brms)
library(rstan)
library(sessioninfo)

# Load model comparison functions
#source("helper/lrtests.R")

# Deal with package priority issues
select <- dplyr::select

theme_set(theme_bw(base_size = 10))
options("future" = T)
# knitr::opts_chunk$set(cache = TRUE)

# print(sessionInfo()) #listing all info about R and packages info
session_info(pkgs = "!attached", to_file = "03b_session_log.txt")

```

# Load dataset and set up contrasts

```{r}
# Read data ======================================
col_types <- cols(
  labid = col_factor(),
  subid = col_factor(),
  subid_unique = col_factor(),
  CDI.form = col_factor(),
  CDI.nwords = col_integer(),
  CDI.prop = col_number(),
  CDI.agerange = col_factor(),
  CDI.agedays = col_integer(),
  CDI.agemin = col_integer(),
  CDI.agemax = col_integer(),
  vocab_nwords = col_integer(),
  standardized.score.CDI = col_character(),
  standardized.score.CDI.num = col_number(),
  IDS_pref = col_number(),
  z.IDS_pref = col_number(),
  language = col_factor(),
  language_zone = col_factor(),
  CDI.error = col_logical(),
  Notes = col_character(),
  trial_order = col_factor(),
  method = col_factor(),
  age_days = col_integer(),
  age_mo = col_number(),
  age_group = col_factor(),
  nae = col_logical(),
  gender = col_factor(),
  second_session = col_logical()
)
data.total <- read_csv("data/02b_processed.csv", col_types = col_types)
# TODO: add saved results
```

## Contrast Setups

We need `gender` as an effect-coded factor, and `method` as a deviation-coded factor. This is achieved in R by using the `contr.sum()` function with the number of levels for each factor. Notably, when subsetting the UK sample, only two levels of `method` out of the three in total were left. 

```{r contrasts}
# Set contrasts on the total dataset =============
contrasts(data.total$gender) <- contr.sum(2)
contrasts(data.total$method) <- contr.sum(3)

# Create sub-datasets, with contrasts ============
## NAE
data.nae <- data.total %>%
  subset(language_zone == "NAE") %>%
  droplevels()
contrasts(data.nae$gender) <- contr.sum(2)
contrasts(data.nae$method) <- contr.sum(3)

## UK (combined-age and separate 18/24 months)

data.uk <- data.total %>%
  subset(language_zone == "British") %>%
  droplevels()
contrasts(data.uk$gender) <- contr.sum(2)
contrasts(data.uk$method) <- contr.sum(2) # note that UK sample has only 2 levels, so sum of zero contrasts set to 2 levels

# Create dataset with British and NAE only
data.uk_nae <- data.total %>%
  subset(language_zone %in% c("British", "NAE")) %>%
  mutate(
    CDI.agemin = ifelse(language_zone == "NAE",
      CDI.agemin + round(.5 * 365.25 / 12),
      CDI.agemin
    ),
    CDI.agemax = ifelse(language_zone == "NAE",
      CDI.agemax - round(.5 * 365.25 / 12),
      CDI.agemax
    )
  ) %>%
  subset(!(CDI.agedays < CDI.agemin | CDI.agedays > CDI.agemax)) %>%
  droplevels()
# Create contrasts for analysis
contrasts(data.uk_nae$gender) <- contr.sum(2)
contrasts(data.uk_nae$method) <- contr.sum(3)
contrasts(data.uk_nae$language_zone) <- contr.sum(2)
```

# Bayesian Statistics

Given that the `frequentist#' tests above did not discover any expected significant effects, we now run a Bayesian analysis, specifically Bayes factors (or $b$-values) from model comparisons, to determine whether our data provide evidence for a true null effect or merely fail to provide evidence either way.

We first run models on the separate UK and NAE samples.

First of all, we need to set up the priors for the Bayeisan analysis. In the RR, we said that we would use a Cauchy distribution for priors over fixed effects with normally distributed priors over random effect parameters."

## Set up priors

Set up a really weak and not too informative prior
```{r weak prior}
prior <- c(
  set_prior("normal(10, 1)", class = "Intercept"), # we can't have negative CDI
  set_prior("cauchy(0, 1)", class = "b"), # all IVs have the same prior..
  set_prior("normal(0, 1)", class = "sd")
)

prior_combined_mod <- c(
  set_prior("normal(10, 1)", class = "Intercept"),
  set_prior("cauchy(0, 1)", class = "b"), # all IVs have the same prior..
  set_prior("normal(0, 1)", class = "sd"),
  set_prior("lkj(2)", class = "L")
) # adding this as there is a random slope and we need a correlation matrix prior between slope and intercept
```

Set up another prior that mainly changed the mean for the fixed effect and see if there is big difference? 
Note: numbers changed but not to the extent that change BF conclusion.
```{r weak prior test sensitivity}
prior_big <- c(
  set_prior("normal(10, 1)", class = "Intercept"), # we can't have negative CDI
  set_prior("cauchy(10, 5)", class = "b"), # all IVs have the same prior..
  set_prior("normal(0, 1)", class = "sd")
)

prior_combined_big <- c(
  set_prior("normal(10, 1)", class = "Intercept"),
  set_prior("cauchy(10, 5)", class = "b"), # all IVs have the same prior..
  set_prior("normal(0, 1)", class = "sd"),
  set_prior("lkj(2)", class = "L")
) # adding this as there is a random slope and we need a correlation matrix prior between slope and intercept
```


We also try more informative prior.

## NAE info prior
```{r informative NAE prior}
prior_nae_1 <- c(
  set_prior("normal(50, 10)", class = "Intercept"), # note that nae's DV is percentile score, but I will assume at 18 months, the average is at 50 percentile
  set_prior("cauchy(5, 1)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 5 % percentile increase?
  set_prior("cauchy(-15, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males
  set_prior("cauchy(0, 1)", class = "b", coef = "z.IDS_pref"), # weak prior as don't know the effect of IDS preference to CDI score
  set_prior("cauchy(0, 1)", class = "b", coef = "method1"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 1)", class = "b", coef = "method2"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
  set_prior("normal(0, 1)", class = "sd")
)

prior_nae_null_1 <- c(
  set_prior("normal(50, 10)", class = "Intercept"), # note that nae's DV is percentile score, but I will assume at 18 months, the average is at 50 percentile
  set_prior("cauchy(5, 1)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 5 % percentile increase?
  set_prior("cauchy(-15, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males
  set_prior("cauchy(0, 1)", class = "b", coef = "method1"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 1)", class = "b", coef = "method2"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
  set_prior("normal(0, 1)", class = "sd")
)

prior_nae_2 <- c(
  set_prior("normal(50, 10)", class = "Intercept"), # note that nae's DV is percentile score, but I will assume at 18 months, the average is at 50 percentile
  set_prior("cauchy(5, 1)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 5 % percentile increase?
  set_prior("cauchy(-15, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males
  set_prior("cauchy(0, 1)", class = "b", coef = "z.IDS_pref"), # weak prior as don't know the effect of IDS preference to CDI score
  set_prior("cauchy(0, 1)", class = "b", coef = "method1"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 1)", class = "b", coef = "method2"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  # note all interaction priors are weak as we don't really have much info to form a prior
  set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
  set_prior("cauchy(0, 1)", class = "b", coef = "CDI.z_age_months:z.IDS_pref"),
  set_prior("cauchy(0, 1)", class = "b", coef = "method1:z.IDS_pref"),
  set_prior("cauchy(0, 1)", class = "b", coef = "method2:z.IDS_pref"),
  set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months:z.IDS_pref"),
  set_prior("normal(0, 1)", class = "sd")
)

prior_nae_null_2 <- c(
  set_prior("normal(50, 10)", class = "Intercept"), # note that nae's DV is percentile score, but I will assume at 18 months, the average is at 50 percentile
  set_prior("cauchy(5, 1)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 5 % percentile increase?
  set_prior("cauchy(-15, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males
  set_prior("cauchy(0, 1)", class = "b", coef = "z.IDS_pref"), # weak prior as don't know the effect of IDS preference to CDI score
  set_prior("cauchy(0, 1)", class = "b", coef = "method1"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 1)", class = "b", coef = "method2"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  # note all interaction priors are weak as we don't really have much info to form a prior
  set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
  set_prior("cauchy(0, 1)", class = "b", coef = "CDI.z_age_months:z.IDS_pref"),
  set_prior("cauchy(0, 1)", class = "b", coef = "method1:z.IDS_pref"),
  set_prior("cauchy(0, 1)", class = "b", coef = "method2:z.IDS_pref"),
  set_prior("normal(0, 1)", class = "sd")
)
```


## UK info prior
```{r}
prior_uk_1 <- c(
  set_prior("normal(40, 20)", class = "Intercept"), # note that UK's DV is raw score, based on wordbank (oxford CDI), the average (50 percentile) is around 41.6
  set_prior("cauchy(15, 1)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 15 words increase (proxy from wordbank)
  set_prior("cauchy(-10, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males
  set_prior("cauchy(0, 1)", class = "b", coef = "z.IDS_pref"), # weak prior as don't know the effect of IDS preference to CDI score
  set_prior("cauchy(0, 1)", class = "b", coef = "method1"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
  set_prior("normal(0, 1)", class = "sd")
)

prior_uk_null_1 <- c(
  set_prior("normal(40, 20)", class = "Intercept"), # note that UK's DV is raw score, based on wordbank (oxford CDI), the average (50 percentile) is around 41.6
  set_prior("cauchy(15, 1)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 15 words increase (proxy from wordbank)
  set_prior("cauchy(-10, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males
  set_prior("cauchy(0, 1)", class = "b", coef = "method1"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
  set_prior("normal(0, 1)", class = "sd")
)

prior_uk_2 <- c(
  set_prior("normal(40, 20)", class = "Intercept"), # note that UK's DV is raw score, based on wordbank (oxford CDI), the average (50 percentile) is around 41.6
  set_prior("cauchy(15, 1)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 15 words increase (proxy from wordbank)
  set_prior("cauchy(-10, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males
  set_prior("cauchy(0, 1)", class = "b", coef = "z.IDS_pref"), # weak prior as don't know the effect of IDS preference to CDI score
  set_prior("cauchy(0, 1)", class = "b", coef = "method1"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
  # note all interaction priors are weak as we don't really have much info to form a prior
  set_prior("cauchy(0, 1)", class = "b", coef = "CDI.z_age_months:z.IDS_pref"),
  set_prior("cauchy(0, 1)", class = "b", coef = "method1:z.IDS_pref"),
  set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months:z.IDS_pref"),
  set_prior("normal(0, 1)", class = "sd")
)

prior_uk_null_2 <- c(
  set_prior("normal(40, 20)", class = "Intercept"), # note that UK's DV is raw score, based on wordbank (oxford CDI), the average (50 percentile) is around 41.6
  set_prior("cauchy(15, 1)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 15 words increase (proxy from wordbank)
  set_prior("cauchy(-10, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males
  set_prior("cauchy(0, 1)", class = "b", coef = "z.IDS_pref"), # weak prior as don't know the effect of IDS preference to CDI score
  set_prior("cauchy(0, 1)", class = "b", coef = "method1"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 1)", class = "b", coef = "z_age_months"),
  # note all interaction priors are weak as we don't really have much info to form a prior
  set_prior("cauchy(0, 1)", class = "b", coef = "CDI.z_age_months:z.IDS_pref"),
  set_prior("cauchy(0, 1)", class = "b", coef = "method1:z.IDS_pref"),
  set_prior("normal(0, 1)", class = "sd")
)
```

## Combined info prior
```{r}

# Note that NAE CDI WS total score is 680 and UK CDI total score is 416.

# NAE average scores for 50th percentile is 76.3, so it is 76.3/680 ~ 11% and UK CDI average scores for 50th percentile is 41.6, so it is 41.6/416 ~ 10%

prior_combined_full_info <- c(
  set_prior("normal(0.1, 0.05)", class = "Intercept"), # given the above info, we expect on average, proportion score is 10%
  set_prior("cauchy(0.03, 0.01)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 15 words increase in UK cdi and lead to 28 words increase in NAE (proxy from wordbank). In general, 15/416 = 3% and 28/680 = 4%
  set_prior("cauchy(-0.03, 0.01)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males, 11/416 ~ 2.6% and Wordbank NAE suggested females knows approximately 27 words than males, 27/680 ~ 3.9%
  set_prior("cauchy(0, 0.01)", class = "b", coef = "language_zone1"), # weak prior as don't know the effect of language zone to CDI score
  set_prior("cauchy(0, 0.01)", class = "b", coef = "z.IDS_pref"), # weak prior as don't know the effect of IDS preference to CDI score
  set_prior("cauchy(0, 0.01)", class = "b", coef = "method1"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 0.01)", class = "b", coef = "method2"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 0.01)", class = "b", coef = "z_age_months"),
  # note all interaction priors are weak as we don't really have much info to form a prior
  set_prior("cauchy(0, 0.01)", class = "b", coef = "CDI.z_age_months:z.IDS_pref"),
  set_prior("cauchy(0, 0.01)", class = "b", coef = "language_zone1:z.IDS_pref"),
  set_prior("cauchy(0, 0.01)", class = "b", coef = "method1:z.IDS_pref"),
  set_prior("cauchy(0, 0.01)", class = "b", coef = "method2:z.IDS_pref"),
  set_prior("cauchy(0, 0.01)", class = "b", coef = "z_age_months:z.IDS_pref"),
  set_prior("normal(0, 0.01)", class = "sd")#,
  #set_prior("lkj(2)", class = "L") #LKJ(2) corr is removed as there is no random slope in this model anymore
) # adding this as there is a random slope and we need a correlation matrix prior between slope and intercept

prior_combined_null_info <- c(
  set_prior("normal(0.1, 0.05)", class = "Intercept"), # given the above info, we expect on average, proportion score is 10%
  set_prior("cauchy(0.03, 0.01)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 15 words increase in UK cdi and lead to 28 words increase in NAE (proxy from wordbank). In general, 15/416 = 3% and 28/680 = 4%
  set_prior("cauchy(-0.03, 0.01)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males, 11/416 ~ 2.6% and Wordbank NAE suggested females knows approximately 27 words than males, 27/680 ~ 3.9%
  set_prior("cauchy(0, 0.01)", class = "b", coef = "language_zone1"), # weak prior as don't know the effect of language zone to CDI score
  set_prior("cauchy(0, 0.01)", class = "b", coef = "z.IDS_pref"), # weak prior as don't know the effect of IDS preference to CDI score
  set_prior("cauchy(0, 0.01)", class = "b", coef = "method1"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 0.01)", class = "b", coef = "method2"), # weak prior as methods related to the MB1 study shouldn't influence CDI scores
  set_prior("cauchy(0, 0.01)", class = "b", coef = "z_age_months"),
  # note all interaction priors are weak as we don't really have much info to form a prior
  set_prior("cauchy(0, 0.01)", class = "b", coef = "CDI.z_age_months:z.IDS_pref"),
  set_prior("cauchy(0, 0.01)", class = "b", coef = "method1:z.IDS_pref"),
  set_prior("cauchy(0, 0.01)", class = "b", coef = "method2:z.IDS_pref"),
  set_prior("cauchy(0, 0.01)", class = "b", coef = "z_age_months:z.IDS_pref"),
  set_prior("normal(0, 0.01)", class = "sd")#,
  #set_prior("lkj(2)", class = "L") #LKJ(2) corr is removed as there is no random slope in this model anymore
) # adding this as there is a random slope and we need a correlation matrix prior between slope and intercept
```


In the following analysis, we follow what we have proposed in the RR about calculating the Bayes factor for the three variables of interest:
1) The main effect IDS preference on CDI scores
2) The interaction effect between IDS preference and IDS test age on CDI scores
3) The interaction effect between language_zone (i.e., dialect in RR) and IDS preference on CDI scores

For the first two variables, we run it on the NAE and UK model separately. For the third variable, we run in on the combined UK & NAE model. 

To calcualte Bayes factor for the main effect IDS preference on CDI scores, we need to two models: a full model that contains the main effect of variables of interest, and the other null model that does not contain the main effect of interest. As all the interaction terms in the models contains IDS preference, in the following, we only include main effects in the full and null models for comparisons.

## NAE model (main effect of IDS preference)
```{r}
Sys.setenv(R_FUTURE_RNG_ONMISUSE = "ignore") # this line of code is needed to get rid of warning message due to the "future" package. https://github.com/satijalab/seurat/issues/3622

lmer.bf_full_1.nae <- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  (1 | labid) + (1 | subid_unique),
data = data.nae,
family = gaussian,
prior = prior_nae_1,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 456
)

summary(lmer.bf_full_1.nae)

lmer.bf_null_1.nae <- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
  z_age_months + method +
  (1 | labid) + (1 | subid_unique),
data = data.nae,
family = gaussian,
prior = prior_nae_null_1,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 123
)

summary(lmer.bf_null_1.nae)
```

Calculate the marginal log likelihoods for each hypothese using Bridge sampling
```{r}

marloglik_full_1.nae <- bridge_sampler(lmer.bf_full_1.nae, silent = T)
marloglik_null_1.nae <- bridge_sampler(lmer.bf_null_1.nae, silent = T)
```

Bayes factor
```{r}

BF_full_1.nae <- bayes_factor(marloglik_full_1.nae, marloglik_null_1.nae)
BF_full_1.nae
```

## NAE model (interaction between IDS test age and IDS preference)
```{r}

lmer.bf_full_2.nae <- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  (1 | labid) + (1 | subid_unique),
data = data.nae,
family = gaussian,
prior = prior_nae_2,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 890
)

summary(lmer.bf_full_2.nae)


lmer.bf_null_2.nae <- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months +
  (1 | labid) + (1 | subid_unique),
data = data.nae,
family = gaussian,
prior = prior_nae_null_2,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 102
)

summary(lmer.bf_null_2.nae)
```

Calculate the marginal log likelihoods for each hypotheses using Bridge sampling
```{r}

marloglik_full_2.nae <- bridge_sampler(lmer.bf_full_2.nae, silent = T)
marloglik_null_2.nae <- bridge_sampler(lmer.bf_null_2.nae, silent = T)
```

Bayes factor
```{r}

BF_full_2.nae <- bayes_factor(marloglik_full_2.nae, marloglik_null_2.nae)
BF_full_2.nae
```

## UK model (main effect of IDS preference)
```{r}

lmer.bf_full_1.uk <- brm(vocab_nwords ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  #(1 | labid) + 
    (1 | subid_unique),
data = data.uk,
family = gaussian,
prior = prior_uk_1,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 213
)

summary(lmer.bf_full_1.uk)


lmer.bf_null_1.uk <- brm(vocab_nwords ~ CDI.z_age_months + gender +
  z_age_months + method +
  #(1 | labid) + 
    (1 | subid_unique),
data = data.uk,
family = gaussian,
prior = prior_uk_null_1,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 312
)

summary(lmer.bf_null_1.uk)
```

Calculate the marginal log likelihoods for each hypothese using Bridge sampling
```{r}
marloglik_full_1.uk <- bridge_sampler(lmer.bf_full_1.uk, silent = T)
marloglik_null_1.uk <- bridge_sampler(lmer.bf_null_1.uk, silent = T)
```

Bayes factor
```{r}
BF_full_1.uk <- bayes_factor(marloglik_full_1.uk, marloglik_null_1.uk)
BF_full_1.uk
```

## UK model (interaction between IDS test age and IDS preference)
```{r}


lmer.bf_full_2.uk <- brm(vocab_nwords ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  #(1 | labid) + 
    (1 | subid_unique),
data = data.uk,
family = gaussian,
prior = prior_uk_2,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 546
)

summary(lmer.bf_full_2.uk)


lmer.bf_null_2.uk <- brm(vocab_nwords ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + 
  #(1 | labid) + 
    (1 | subid_unique),
data = data.uk,
family = gaussian,
prior = prior_uk_null_2,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 689
)

summary(lmer.bf_null_2.uk)
```

### Calculate the marginal log likelihoods for each hypothese using Bridge sampling

```{r}
marloglik_full_2.uk <- bridge_sampler(lmer.bf_full_2.uk, silent = T)
marloglik_null_2.uk <- bridge_sampler(lmer.bf_null_2.uk, silent = T)
```

### Bayes factor

```{r}
BF_full_2.uk <- bayes_factor(marloglik_full_2.uk, marloglik_null_2.uk)
BF_full_2.uk
```

## Combined model (interaction between lanaguage zone and IDS preference)
```{r Bayesian_combined Bayes Factor analysis}

lmer.bf_full_3.combined <- brm(CDI.prop ~ CDI.z_age_months + language_zone + gender +
  z_age_months + method + z.IDS_pref + z.IDS_pref:language_zone +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  (1 | labid) + (1 | subid_unique),
data = data.uk_nae,
family = gaussian,
prior = prior_combined_full_info,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .9999),
seed = 100
)

summary(lmer.bf_full_3.combined)


lmer.bf_null_3.combined <- brm(CDI.prop ~ CDI.z_age_months + language_zone + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  (1 | labid) + (1 | subid_unique),
data = data.uk_nae,
family = gaussian,
prior = prior_combined_null_info,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .9999),
seed = 200
)

summary(lmer.bf_null_3.combined)
```

### Calculate the marginal log likelihoods for each hypothese using Bridge sampling

```{r}
marloglik_full_3.combined <- bridge_sampler(lmer.bf_full_3.combined, silent = T)
marloglik_null_3.combined <- bridge_sampler(lmer.bf_null_3.combined, silent = T)
```

### Bayes factor

```{r}
BF_full_3.combined <- bayes_factor(marloglik_full_3.combined, marloglik_null_3.combined)
BF_full_3.combined
```


# Robustness checks on Bayesian stats

```{r}
make_prior_nae <- function(scale){
  p <- str_c("cauchy(0, ",as.character(scale),")")
  prior_nae <- c(
  set_prior("normal(50, 10)", class = "Intercept"), # note that nae's DV is percentile score, but I will assume at 18 months, the average is at 50 percentile
  set_prior("cauchy(5, 1)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 5 % percentile increase?
  set_prior("cauchy(-15, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males
  set_prior(p, class = "b"), # everything else 
  set_prior("normal(0, 1)", class = "sd")
)
  return(prior_nae)
}

make_prior_uk <- function(scale){
    p <- str_c("cauchy(0, ",as.character(scale),")")
  prior_uk <- c(
  set_prior("normal(40, 20)", class = "Intercept"), # note that UK's DV is raw score, based on wordbank (oxford CDI), the average (50 percentile) is around 41.6
  set_prior("cauchy(15, 1)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 15 words increase (proxy from wordbank)
  set_prior("cauchy(-10, 5)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males
  set_prior(p, class = "b"), # everything else
  set_prior("normal(0, 1)", class = "sd")
)
  return(prior_uk)
}

make_prior_combined <- function(scale){
      p <- str_c("cauchy(0, ",as.character(scale),")")
    prior_combined <- c(
      set_prior("normal(0.1, 0.05)", class = "Intercept"), # given the above info, we expect on average, proportion score is 10%
      set_prior("cauchy(0.03, 0.01)", class = "b", coef = "CDI.z_age_months"), # assume increase in 1 month may lead to 15 words increase in UK cdi and lead to 28 words increase in NAE (proxy from wordbank). In general, 15/416 = 3% and 28/680 = 4%
      set_prior("cauchy(-0.03, 0.01)", class = "b", coef = "gender1"), # female is the baseline, normally females know more words than males. Wordbank oxford CDI suggested at 18 months (50 percentile), female knows approximately 11 words more than males, 11/416 ~ 2.6% and Wordbank NAE suggested females knows approximately 27 words than males, 27/680 ~ 3.9%
      set_prior(p, class = "b"), 
      set_prior("normal(0, 0.01)", class = "sd"))
    return(prior_combined)
}

```

```{r robustness_nae_main}

do_NAE_main_model <- function(scale, save_full, save_null){
full <- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  (1 | labid) + (1 | subid_unique),
data = data.nae,
family = gaussian,
prior = make_prior_nae(scale),
file=save_full,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 456
)

null <- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
  z_age_months + method +
  (1 | labid) + (1 | subid_unique),
data = data.nae,
family = gaussian,
prior = make_prior_nae(scale),
file=save_null,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 123
)

    
marloglik_full <- bridge_sampler(full, silent = T)
marloglik_null <- bridge_sampler(null, silent = T)

BF <- bayes_factor(marloglik_full, marloglik_null)
return(BF)
}
```


```{r robustness_nae_interact}

do_NAE_interact_model <- function(scale, save_full, save_null){
 full<- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  (1 | labid) + (1 | subid_unique),
data = data.nae,
family = gaussian,
prior = make_prior_nae(scale),
file=save_full,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 890
)

null<- brm(standardized.score.CDI.num ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months +
  (1 | labid) + (1 | subid_unique),
data = data.nae,
family = gaussian,
prior = make_prior_nae(scale),
file=save_null,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 102
)

marloglik_full <- bridge_sampler(full, silent = T)
marloglik_null <- bridge_sampler(null, silent = T)

BF <- bayes_factor(marloglik_full, marloglik_null)
return(BF)
}
```

```{r robustness_uk_main}

do_UK_main_model <- function(scale, save_full, save_null){
full <- brm(vocab_nwords ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  #(1 | labid) + 
    (1 | subid_unique),
data = data.uk,
family = gaussian,
prior = make_prior_uk(scale),
file=save_full,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 456
)

null <- brm(vocab_nwords ~ CDI.z_age_months + gender +
  z_age_months + method +
  #(1 | labid) + 
    (1 | subid_unique),
data = data.uk,
family = gaussian,
prior = make_prior_uk(scale),
file=save_null,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 123
)

    
marloglik_full <- bridge_sampler(full, silent = T)
marloglik_null <- bridge_sampler(null, silent = T)

BF <- bayes_factor(marloglik_full, marloglik_null)
return(BF)
}
```

```{r robustness_uk_interact}

do_UK_interact_model <- function(scale, save_full, save_null){
full <- brm(vocab_nwords ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  #(1 | labid) + 
    (1 | subid_unique),
data = data.uk,
family = gaussian,
prior = make_prior_uk(scale),
file=save_full,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 890
)

null<-  brm(vocab_nwords ~ CDI.z_age_months + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + 
  #(1 | labid) + 
    (1 | subid_unique),
data = data.uk,
family = gaussian,
prior = make_prior_uk(scale),
file=save_null,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .95),
seed = 102
)

marloglik_full <- bridge_sampler(full, silent = T)
marloglik_null <- bridge_sampler(null, silent = T)

BF <- bayes_factor(marloglik_full, marloglik_null)
return(BF)
}
```

```{r robustness_combined_dialect_interact}

do_combined_model <- function(scale, save_full, save_null){
 full<- brm(CDI.prop ~ CDI.z_age_months + language_zone + gender +
  z_age_months + method + z.IDS_pref + z.IDS_pref:language_zone +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  (1 | labid) + (1 | subid_unique),
data = data.uk_nae,
family = gaussian,
prior = make_prior_combined(scale),
file=save_full,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .9999),
seed = 100
)


null <- brm(CDI.prop ~ CDI.z_age_months + language_zone + gender +
  z_age_months + method + z.IDS_pref +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  (1 | labid) + (1 | subid_unique),
data = data.uk_nae,
family = gaussian,
prior = make_prior_combined(scale),
file=save_null,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .9999),
seed = 200
)

marloglik_full <- bridge_sampler(full, silent = T)
marloglik_null <- bridge_sampler(null, silent = T)

BF <- bayes_factor(marloglik_full, marloglik_null)
return(BF)
}
```

```{r}
b <- do_NAE_main_model(10, "mods/NAE_main_full_10.rds", "mods/NAE_main_null_10.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.37045
a <- do_NAE_main_model(1, "mods/NAE_main_full_1.rds", "mods/NAE_main_null_1.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.86690
c <- do_NAE_main_model(.1, "mods/NAE_main_full_point1.rds", "mods/NAE_main_null_point1.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.86831
p <- do_NAE_main_model(.001, "mods/NAE_main_full_point001.rds", "mods/NAE_main_null_point001.rds")
#Estimated Bayes factor in favor of x1 over x2: 1.57793

e <- do_NAE_interact_model(10, "mods/NAE_interact_full_10.rds", "mods/NAE_interact_null_10.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.33237
d <- do_NAE_interact_model(1, "mods/NAE_interact_full_1.rds", "mods/NAE_interact_null_1.rds")
#Estimated Bayes factor in favor of x1 over x2: 1.03461
f <- do_NAE_interact_model(.1, "mods/NAE_interact_full_point1.rds", "mods/NAE_interact_null_point1.rds")
#Estimated Bayes factor in favor of x1 over x2: 1.03096
t <- do_NAE_interact_model(.01, "mods/NAE_interact_full_point01.rds", "mods/NAE_interact_null_point01.rds")
#Estimated Bayes factor in favor of x1 over x2: 1.19879
q <- do_NAE_interact_model(.001, "mods/NAE_interact_full_point001.rds", "mods/NAE_interact_null_point001.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.63298

v <- do_UK_main_model(100, "mods/UK_main_full_100.rds", "mods/UK_main_null_100.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.20513

h <- do_UK_main_model(10, "mods/UK_main_full_10.rds", "mods/UK_main_null_10.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.77336
w <- do_UK_main_model(5, "mods/UK_main_full_5.rds", "mods/UK_main_null_5.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.89392
g <- do_UK_main_model(1, "mods/UK_main_full_1.rds", "mods/UK_main_null_1.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.96754
x <- do_UK_main_model(.5, "mods/UK_main_full_point5.rds", "mods/UK_main_null_point5.rds")
#Estimated Bayes factor in favor of x1 over x2: 1.00724
i <- do_UK_main_model(.1, "mods/UK_main_full_point1.rds", "mods/UK_main_null_point1.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.90475
u <- do_UK_main_model(.01, "mods/UK_main_full_point01.rds", "mods/UK_main_null_point01.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.96621

z <-  do_UK_interact_model(50, "mods/UK_interact_50.rds", "mods/UK_interact_null_50.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.17077

k <- do_UK_interact_model(10, "mods/UK_interact_full_10.rds", "mods/UK_interact_null_10.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.47832
j <- do_UK_interact_model(1, "mods/UK_interact_full_1.rds", "mods/UK_interact_null_1.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.97005
l <- do_UK_interact_model(.1, "mods/UK_interact_full_point1.rds", "mods/UK_interact_null_point1.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.98716
y <- do_UK_interact_model(.01, "mods/UK_interact_full_point01.rds", "mods/UK_interact_null_point01.rds")
#Estimated Bayes factor in favor of x1 over x2: 1.14301

n <- do_combined_model(.1, "mods/combined_full_point1.rds", "mods/combined_null_point1.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.21141
m <- do_combined_model(.01, "mods/combined_full_point01.rds", "mods/combined_null_point01.rds")
# Estimated Bayes factor in favor of x1 over x2: 0.28282

aa <- do_combined_model(.007, "mods/combined_full_point007.rds", "mods/combined_null_point007.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.50509
r <- do_combined_model(.005, "mods/combined_full_point005.rds", "mods/combined_null_point005.rds")
#Estimated Bayes factor in favor of x1 over x2: 1.22465
o <- do_combined_model(.001, "mods/combined_full_point001.rds", "mods/combined_null_point001.rds")
#Estimated Bayes factor in favor of x1 over x2: 1.14791
bb <- do_combined_model(.0005, "mods/combined_full_point0005.rds", "mods/combined_null_point0005.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.42625

s <- do_combined_model(.0001, "mods/combined_full_point0001.rds", "mods/combined_null_point0001.rds")
#Estimated Bayes factor in favor of x1 over x2: 2646871512278566400.00000
```

## Model + robustness 

### Combined on main effect
```{r robustness_combined_main}

do_combined_model_main <- function(scale, save_full, save_null){
 full<- brm(CDI.prop ~ CDI.z_age_months + language_zone + gender +
  z_age_months + method + z.IDS_pref +
  (1 | labid) + (1 | subid_unique),
data = data.uk_nae,
family = gaussian,
prior = make_prior_combined(scale),
file=save_full,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .9999),
seed = 100
)


null <- brm(CDI.prop ~ CDI.z_age_months + language_zone + gender +
  z_age_months + method + 
  (1 | labid) + (1 | subid_unique),
data = data.uk_nae,
family = gaussian,
prior = make_prior_combined(scale),
file=save_null,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .9999),
seed = 200
)

marloglik_full <- bridge_sampler(full, silent = T)
marloglik_null <- bridge_sampler(null, silent = T)

BF <- bayes_factor(marloglik_full, marloglik_null)
return(BF)
}
```

### Combined on interaction

```{r robustness_combined_interact}

do_combined_model_interact <- function(scale, save_full, save_null){
 full<- brm(CDI.prop ~ CDI.z_age_months + language_zone + gender +
  z_age_months + method + z.IDS_pref + z.IDS_pref:language_zone +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + z.IDS_pref:z_age_months +
  (1 | labid) + (1 | subid_unique),
data = data.uk_nae,
family = gaussian,
prior = make_prior_combined(scale),
file=save_full,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .9999),
seed = 100
)


null <- brm(CDI.prop ~ CDI.z_age_months + language_zone + gender +
  z_age_months + method + z.IDS_pref + z.IDS_pref:language_zone +
  z.IDS_pref:method + z.IDS_pref:CDI.z_age_months + 
  (1 | labid) + (1 | subid_unique),
data = data.uk_nae,
family = gaussian,
prior = make_prior_combined(scale),
file=save_null,
iter = 4000,
warmup = 2000,
chains = 4,
save_pars = save_pars(all = T),
control = list(adapt_delta = .9999),
seed = 200
)

marloglik_full <- bridge_sampler(full, silent = T)
marloglik_null <- bridge_sampler(null, silent = T)

BF <- bayes_factor(marloglik_full, marloglik_null)
return(BF)
}
```

```{r}
# with default value
# main_combined <- do_combined_model_main(.01, "mods/combined_main_full_point01.rds", "mods/combined_main_null_point01.rds")
# # Estimated Bayes factor in favor of x1 over x2: 0.74997
# 
# interact_combined <- do_combined_model_interact(.01, "mods/combined_interact_full_point01.rds", "mods/combined_interact_null_point01.rds")
# # Estimated Bayes factor in favor of x1 over x2: 0.34544
# 
# ## for robustness
# 
# bb <- do_combined_model_main(.001, "mods/combined_main_full_point001.rds", "mods/combined_main_null_point001.rds")
# #Estimated Bayes factor in favor of x1 over x2: 0.88511
# 
# cc <- do_combined_model_main(.1, "mods/combined_main_full_point1.rds", "mods/combined_main_null_point1.rds")
# #Estimated Bayes factor in favor of x1 over x2: 0.27974

#ff <- do_combined_model_main(.05, "mods/combined_main_full_point05.rds", "mods/combined_main_null_point05.rds")
#Estimated Bayes factor in favor of x1 over x2: 4.20642
#xx <- do_combined_model_main(.03, "mods/combined_main_full_point03.rds", "mods/combined_main_null_point03.rds")
#Estimated Bayes factor in favor of x1 over x2: 3.01790

# dd <- do_combined_model_interact(.001, "mods/combined_interact_full_point001.rds", "mods/combined_interact_null_point001.rds")
# #Estimated Bayes factor in favor of x1 over x2: 2.70523
# 
# ee <- do_combined_model_interact(.1, "mods/combined_interact_full_point1.rds", "mods/combined_interact_null_point1.rds")
# #Estimated Bayes factor in favor of x1 over x2: 0.04190

#gg <- do_combined_model_interact(.005, "mods/combined_interact_full_point005.rds", "mods/combined_interact_null_point005.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.88843

#yy <- do_combined_model_interact(.05, "mods/combined_interact_full_point05.rds", "mods/combined_interact_null_point05.rds")
#Estimated Bayes factor in favor of x1 over x2: 0.49613
```
