==> 04_compute_percentiles_mb_cdi.Rmd <==
---
title: "Processing manybabies cdi project"
author: "Natalia (07/2021) Altered by: Luis (02/2022) and Joscelin (09/16/2022)"
date: '`r format(Sys.time(), "%a %b %d %X %Y")`'
output: 
  html_document:
    toc: true
    toc_depth: 5
    number_sections: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load Libraries

```{r libraries}
library(tidyverse)
library(sessioninfo)

session_info(pkgs = "!attached", to_file = "04_session_log.txt")
```

## Load data

```{r load data}
rm(list = ls())

# Load data for each lab's vocabulary scores to compute percentiles
d1 <- read.csv("data/02b_processed.csv")

# Examine the outcomes of variables
summary(d1)
```

## Load data for individual type

### Korean CDI percentiles

```{r korean}
# Load computed percentile rankings computed from norms of vocabulary scores 
# for the CDI instrument that was used in that country
korean <- read.delim("data/cdi_percentile/korean.txt", header = F)

# Rename the columns to match percentile ranking
colnames(korean) <- c(1:99)

# Get the minimum age for the CDI instrument
min_age_CDI1 <- which(korean[, 1] != 0)[1]

# Get the maximal age for the CDI instrument
max_age_CDI1 <- nrow(korean)

# Add a column that uses age in days from the computed growth curves
korean$Age <- c(1:max_age_CDI1)

# The value in each cell indicates the number of words a child of a given age at a given percentile ranking would know and produce.
# The columns indicate the percentile ranking and the rows indicate the age in days.
# There are zeroes for all the rows that the age in days is outside of the validated age for the CDI tool.
# Please scroll until you move past the zeroes and onto the actual vocabulary values assigned to each percentile ranking for each age in days.
```

### Oxford CDI percentiles

```{r oxford}
# Load computed percentile rankings computed from norms of vocabulary scores 
# for the CDI instrument that was used in that country
oxford <- read.delim("data/cdi_percentile/oxford.txt", header = F)

# Rename the columns to match percentile ranking
colnames(oxford) <- c(1:99)

# Get the minimum age for the CDI instrument
min_age_CDI1 <- which(oxford[, 1] != 0)[1]

# Get the maximal age for the CDI instrument
max_age_CDI1 <- nrow(oxford)

# Add a column that uses age in days from the computed growth curves
oxford$Age <- c(1:max_age_CDI1)

# The value in each cell indicates the number of words a child of a given age at a given percentile ranking would know and produce.
# The columns indicate the percentile ranking and the rows indicate the age in days.
# There are zeroes for all the rows that the age in days is outside of the validated age for the CDI tool.
# Please scroll until you move past the zeroes and onto the actual vocabulary values assigned to each percentile ranking for each age in days.
```

### Web CDI percentiles

```{r web, echo=TRUE, include=FALSE}
# Load computed percentile rankings computed from norms of vocabulary scores 
# for the CDI instrument that was used in that country
web_cdi <- read.delim("data/cdi_percentile/web_cdi.txt", header = F)

# Rename the columns to match percentile ranking
colnames(web_cdi) <- c(1:99)

# Get the minimum age for the CDI instrument
min_age_CDI1 <- which(web_cdi[, 1] != 0)[1]

# Get the maximal age for the CDI instrument
max_age_CDI1 <- nrow(web_cdi)

# Add a column that uses age in days from the computed growth curves
web_cdi$Age <- c(1:max_age_CDI1)

# The value in each cell indicates the number of words a child of a given age at a given percentile ranking would know and produce.
# The columns indicate the percentile ranking and the rows indicate the age in days.
# There are zeroes for all the rows that the age in days is outside of the validated age for the CDI tool.
# Please scroll until you move past the zeroes and onto the actual vocabulary values assigned to each percentile ranking for each age in days.
```

### Norwegian CDI percentiles

```{r norwegian}
# Load computed percentile rankings computed from norms of vocabulary scores 
# for the CDI instrument that was used in that country
norwegian <- read.delim("data/cdi_percentile/norwegian.txt", header = F)

# Rename the columns to match percentile ranking
colnames(norwegian) <- c(1:99)

# Get the minimum age for the CDI instrument
min_age_CDI1 <- which(norwegian[, 1] != 0)[1]

# Get the maximal age for the CDI instrument
max_age_CDI1 <- nrow(norwegian)

# Add a column that uses age in days. The Norwegian data was originally collected in days.
# Growth curves weren't used
norwegian$Age <- c(1:max_age_CDI1)

# The value in each cell indicates the number of words a child of a given age at a given percentile ranking would know and produce.
# The columns indicate the percentile ranking and the rows indicate the age in days.
# There are zeroes for all the rows that the age in days is outside of the validated age for the CDI tool.
# Please scroll until you move past the zeroes and onto the actual vocabulary values assigned to each percentile ranking for each age in days.
```

### German CDI percentiles

```{r german}
# Load computed percentile rankings computed from norms of vocabulary scores 
# for the CDI instrument that was used in that country
german <- read.delim("data/cdi_percentile/german.txt", header = F)

# Rename the columns to match percentile ranking
colnames(german) <- c(1:99)

# Get the minimum age for the CDI instrument
min_age_CDI1 <- which(german[, 1] != 0)[1]

# Get the maximal age for the CDI instrument
max_age_CDI1 <- nrow(german)

# Add a column that uses age in days from the computed growth curves
german$Age <- c(1:max_age_CDI1)

# The value in each cell indicates the number of words a child of a given age at a given percentile ranking would know and produce.
# The columns indicate the percentile ranking and the rows indicate the age in days.
# There are zeroes for all the rows that the age in days is outside of the validated age for the CDI tool.
# Please scroll until you move past the zeroes and onto the actual vocabulary values assigned to each percentile ranking for each age in days.
```

### Dutch CDI percentiles

```{r dutch}
# Load computed percentile rankings computed from norms of vocabulary scores 
# for the CDI instrument that was used in that country
dutch <- read.delim("data/cdi_percentile/dutch.txt", header = F)

# Rename the columns to match percentile ranking
colnames(dutch) <- c(1:99)

# Get the minimum age for the CDI instrument
min_age_CDI1 <- which(dutch[, 1] != 0)[1]

# Get the maximal age for the CDI instrument
max_age_CDI1 <- nrow(dutch)

# Add a column that uses age in days from the computed growth curves
dutch$Age <- c(1:max_age_CDI1)

# The value in each cell indicates the number of words a child of a given age at a given percentile ranking would know and produce.
# The columns indicate the percentile ranking and the rows indicate the age in days.
# There are zeroes for all the rows that the age in days is outside of the validated age for the CDI tool.
# Please scroll until you move past the zeroes and onto the actual vocabulary values assigned to each percentile ranking for each age in days.
```

### French CDI percentiles

```{r french}
# Load computed percentile rankings computed from norms of vocabulary scores 
# for the CDI instrument that was used in that country
french <- read.delim("data/cdi_percentile/french.txt", header = F)

# Rename the columns to match percentile ranking
colnames(french) <- c(1:99)

# Get the minimum age for the CDI instrument
min_age_CDI1 <- which(french[, 1] != 0)[1]

# Get the maximal age for the CDI instrument
max_age_CDI1 <- nrow(french)

# Add a column that uses age in days from the computed growth curves
french$Age <- c(1:max_age_CDI1)

# The value in each cell indicates the number of words a child of a given age at a given percentile ranking would know and produce.
# The columns indicate the percentile ranking and the rows indicate the age in days.
# There are zeroes for all the rows that the age in days is outside of the validated age for the CDI tool.
# Please scroll until you move past the zeroes and onto the actual vocabulary values assigned to each percentile ranking for each age in days.
```

## Compute CDI in percentiles for production

```{r Compute CDI in percentiles for production}

# Initiate variables that will be used later in the code chunk
prod_T1 <- c()
perc <- NULL
perc_15plus <- NULL

# Run through each row in d1. Each row corresponds to a single participant from the data the labs collected.
# This loop assigns a percentile ranking from 1 to 99th for each participant based on the norms
# of the instrument that was used in each lab. The aim is to match our participant's age in days to the age in days
# for the norm data. Once this row is found, it compares the value in each cell (which represents the number of words known)
# between our participant and the norms. The column with the closest matching known words from the norms is then assigned
# the percentile ranking for our participant.

for (i in 1:nrow(d1))
{
  tool <- d1[[i, "CDI.form"]]
  # Pick which CDI daily vocabulary score rankings to use based on the language.
  if ((tool == "O-CDI")) {
    perc_CDI1_prod <- oxford
  }

  if ((tool == "FRAKIS")) {
    perc_CDI1_prod <- german
  }

  if ((tool == "webCDI")) {
    perc_CDI1_prod <- web_cdi
  }

  if ((tool == "NorCDI")) {
    perc_CDI1_prod <- norwegian
  }

  if ((tool == "K-CDI")) {
    perc_CDI1_prod <- korean
  }

  if ((tool == "NCDI")) {
    perc_CDI1_prod <- dutch
  }

  if ((tool == "French European")) {
    perc_CDI1_prod <- french
  }

  # Get the minimum age for the CDI instrument
  min_age_CDI1 <- which(perc_CDI1_prod[, 1] != 0)[1]

  # Get the maximal age for the CDI instrument
  max_age_CDI1 <- nrow(perc_CDI1_prod)

  # Get percentiles for production (both for CDI-I[WG] and CDI-II[WS])
  score <- as.numeric(d1[[i, "vocab_nwords"]])

  # Get age in days for the current participant
  age <- as.numeric(d1[[i, "CDI.agedays"]])

  # Account for the differences in the way that countries defined their norming

  # age for the CDI.
  perc_CDI1_prod$Age_15plus <- perc_CDI1_prod$Age + 15

  # If the participant's age is above or below the CDI age, or their score is NA assign their score NA.
  if (age > max_age_CDI1 | age < min_age_CDI1 + 15 | is.na(score)) {
    perc_15plus <- NA
  } else {

    # Get the line number with the age corresponding to the current participant's
    # age.
    age_row <- as.numeric(which(perc_CDI1_prod$Age_15plus == age))

    # Get the number of the column containing the min value between the CDI score and the value in the age corresponding line
    perc_15plus <- as.numeric(which.min(abs(perc_CDI1_prod[age_row, c(1:99)] - score)))

    if (score == 0) {
      # Some infants have a score of 0 even at different ages and this can be perfectly normally depending on their age, rather than assigning them 0 percentiles we assign them the percentile they were divided by 2 and rounded up. This avoids systematic zeros while maintaining the properties of the distribution.
      perc_15plus <- ceiling(perc_15plus / 2)
    }
  }

  # Do not apply the 15 day correction to Norwegian data because the age in days was collected rather than interpolated.
  if (tool == "NorCDI" | tool == "K-CDI") {
    # Adjust for the age limits in the sample-specific norming data
    if (age > max_age_CDI1 | age < min_age_CDI1 | is.na(score)) {
      perc <- NA
    } else {
      # Get the number of line with the corresponding age
      age_row <- as.numeric(which(perc_CDI1_prod$Age == age))
      # Get the number of the column containing the min value between the CDI
      # score and the value in the age corresponding line
      perc <- as.numeric(which.min(abs(perc_CDI1_prod[age_row, c(1:99)] - score)))
      if (score == 0) {
        perc <- ceiling(perc / 2)
      }
    }
    perc_15plus <- perc
  }
  if ((tool == "FROSCH" || tool == "NCDI")) {
    perc <- NA
  }
  # Join all the rows from prod_T1 and perc_15plus to get the percentiles for all the data. Most percentile rankings are in perc_15plus
  prod_T1 <- c(prod_T1, perc_15plus)
}

# Add a column for the CDI production score in percentiles for T1
d1$daily_percentile <- prod_T1


# The Swiss data already had the percentiles computed using the same method as above. This code is to include the correctly computed Swiss
# percentiles into the column daily_percentile for the joint data set. 

# read data set
swiss_data <- read.delim("data/cdi_percentile/zurich_data.txt", header = T)

# Include only the participants where the babies were within the right age to be tested with the CDI tool.
d1[d1$labid == "weltentdeckerzurich", "daily_percentile"] <- swiss_data[swiss_data$CDIagerange == "18","percentile"]

# Write out the joint data set. 
write.table(d1, "data/cdi_percentile/percentiles_manybabies_cdi.txt", 
            sep = "\t", row.names = FALSE, quote = FALSE)
```
